@@ -15,32 +15,39 @@
     "url": "git+https://github.com/andes/app.git"
   },
   "author": "Red TICs",
-  "license": "GPL",
+  "license": "GPL-3.0",
   "bugs": {
     "url": "https://github.com/andes/app/issues"
   },
   "dependencies": {
-    "@andes/auth": "^0.2.4",
-    "@andes/plex": "^1.1.4",
-    "@andes/shared": "^0.4.1",
-    "@angular/common": "2.4.8",
-    "@angular/compiler": "2.4.8",
-    "@angular/core": "2.4.8",
-    "@angular/forms": "2.4.8",
-    "@angular/http": "2.4.8",
-    "@angular/material": "^2.0.0-beta.2",
-    "@angular/platform-browser": "2.4.8",
-    "@angular/platform-browser-dynamic": "2.4.8",
-    "@angular/router": "3.4.8",
-    "angular-confirmation-popover": "^2.1.2",
+    "@andes/auth": "^1.0.0",
+    "@andes/match": "^0.1.9",
+    "@andes/plex": "^2.0.13",
+    "@andes/shared": "^1.0.1",
+    "@angular/animations": "^4.3.1",
+    "@angular/cdk": "^2.0.0-beta.8",
+    "@angular/common": "^4.3.1",
+    "@angular/compiler": "^4.3.1",
+    "@angular/core": "^4.3.1",
+    "@angular/forms": "^4.3.1",
+    "@angular/http": "^4.3.1",
+    "@angular/material": "^2.0.0-beta.8",
+    "@angular/platform-browser": "^4.3.1",
+    "@angular/platform-browser-dynamic": "^4.3.1",
+    "@angular/platform-server": "^4.3.1",
+    "@angular/router": "^4.3.1",
     "angular2-google-maps": "^0.17.0",
+    "bowser": "^1.7.1",
+    "bson": "^1.0.4",
     "core-js": "^2.4.1",
-    "rxjs": "5.1.1",
-    "zone.js": "0.7.7"
+    "enhanced-resolve": "^3.4.1",
+    "ng2-drag-drop": "^1.5.1",
+    "rxjs": "^5.4.2",
+    "zone.js": "^0.8.14"
   },
   "devDependencies": {
-    "@angular/cli": "1.0.0-beta.32.3",
-    "@angular/compiler-cli": "2.4.8",
+    "@angular/cli": "^1.3.2",
+    "@angular/compiler-cli": "^4.3.1",
     "@types/jasmine": "^2.2.30",
     "@types/node": "^6.0.42",
     "codelyzer": "~2.0.0-beta.4",
@@ -49,12 +56,13 @@
     "karma": "~1.4.1",
     "karma-chrome-launcher": "~2.0.0",
     "karma-cli": "~1.0.1",
+    "karma-coverage-istanbul-reporter": "^0.2.0",
     "karma-jasmine": "~1.1.0",
     "karma-jasmine-html-reporter": "^0.2.2",
-    "karma-coverage-istanbul-reporter": "^0.2.0",
     "protractor": "~5.1.0",
     "ts-node": "~2.0.0",
     "tslint": "~4.4.2",
-    "typescript": "~2.0.0"
+    "typescript": "^2.4.2",
+    "webpack": "^3.4.1"
   }
 }
