@@ -2,6 +2,7 @@ import { environment } from './../environments/environment';
 import { Component, OnInit, ModuleWithProviders } from '@angular/core';
 import { Plex } from '@andes/plex';
 import { Server } from '@andes/shared';
+import { Auth } from '@andes/auth';
 
 @Component({
     selector: 'app',
@@ -9,27 +10,73 @@ import { Server } from '@andes/shared';
 })
 
 export class AppComponent {
-    constructor(public plex: Plex, public server: Server) {
+    private initStatusCheck() {
+        if (environment.APIStatusCheck) {
+            setTimeout(() => {
+                this.server.get('/core/status', { params: null, showError: false, showLoader: false })
+                    .finally(() => this.initStatusCheck())
+                    .subscribe(
+                    (data) => this.plex.updateAppStatus(data),
+                    (err) => this.plex.updateAppStatus({ API: 'Error' })
+                    );
+            }, 2000);
+        } else {
+            this.plex.updateAppStatus({ API: 'OK' });
+        }
+    }
+
+    private menuList = [];
+    public checkPermissions(): any {
+        let accessList = [];
+        this.menuList = [];
+
+        if (this.auth.loggedIn()) {
+            this.auth.organizaciones().subscribe(data => {
+                if (data.length > 1) {
+                    this.menuList = [{ label: 'Seleccionar organización', icon: 'home', route: '/selectOrganizacion' }, ...this.menuList];
+                    this.plex.updateMenu(this.menuList);
+                }
+            });
+        }
+        // Cargo el array de permisos
+        if (this.auth.getPermissions('turnos:planificarAgenda:?').length > 0) {
+            accessList.push({ label: 'CITAS: Gestor de Agendas y Turnos', icon: 'calendar', route: '/citas/gestor_agendas' });
+        }
+        if (this.auth.getPermissions('turnos:darTurnos:?').length > 0) {
+            accessList.push({ label: 'CITAS: Punto de Inicio', icon: 'calendar', route: '/puntoInicioTurnos' });
+        }
+        if (this.auth.getPermissions('mpi:?').length > 0) {
+            accessList.push({ label: 'MPI: Indice Maestro de Pacientes', icon: 'account-multiple-outline', route: '/mpi' });
+        }
+
+        if (this.auth.getPermissions('rup:?').length > 0) {
+            accessList.push({ label: 'RUP: Registro Universal de Prestaciones', icon: 'contacts', route: '/rup' });
+        }
+
+        this.menuList.push({ label: 'Página principal', icon: 'home', route: '/inicio' });
+
+        accessList.forEach((permiso) => {
+            this.menuList.push(permiso);
+        });
+        this.menuList.push({ divider: true });
+        this.menuList.push({ label: 'Cerrar Sesión', icon: 'logout', route: '/login' });
+
+        // Actualizamos la lista de menú
+        this.plex.updateMenu(this.menuList);
+        return accessList;
+    }
+
+    constructor(public plex: Plex, public server: Server, public auth: Auth) {
         // Configura server. Debería hacerse desde un provider (http://stackoverflow.com/questions/39033835/angularjs2-preload-server-configuration-before-the-application-starts)
         server.setBaseURL(environment.API);
-     }
 
-    // ngOnInit() {
-        // const items = [
-        //     new MenuItem({ label: 'Inicio', icon: 'creation', route: '/inicio' }),
-        //     new MenuItem({ label: 'Organizacion', icon: 'hospital-building', route: '/organizacion' }),
-        //     new MenuItem({ label: 'Profesional', icon: 'account-circle', route: '/profesional' }),
-        //     new MenuItem({ label: 'Especialidad', icon: 'certificate', route: '/especialidad' }),
-        //     new MenuItem({ label: 'Paciente', icon: 'seat-recline-normal', route: '/paciente' }),
-        //     new MenuItem({ label: 'Espacio Físico', icon: 'view-agenda', route: '/espacio_fisico' }),
-        //     new MenuItem({ label: 'Prestacion', icon: 'blur', route: '/prestacion' }),
-        //     new MenuItem({ label: 'Agendas', icon: 'calendar', route: '/agenda' }),
-        //     new MenuItem({ label: 'Turnos', icon: 'calendar-check', route: '/turnos' }),
-        //     new MenuItem({ label: 'Lista de Espera', icon: 'calendar-check', route: '/listaEspera' }),
-        //     new MenuItem({ label: 'Gestor Agendas', icon: 'calendar-check', route: '/gestor_agendas' }),
-        //     new MenuItem({ label: 'rup Prestaciones', icon: 'calendar-check', route: '/rup' }),
-        //     new MenuItem({ label: 'Tipo de Prestaciones', icon: 'blur', route: '/tipoprestaciones' }),
-        // ];
-        // this.plex.initStaticItems(items);
-    // }
+        // Inicializa el menu
+        this.checkPermissions();
+
+        // Inicializa la vista
+        this.plex.updateTitle('ANDES | Apps Neuquinas de Salud');
+
+        // Inicializa el chequeo de conectividad
+        this.initStatusCheck();
+    }
 }
