@@ -1,6 +1,3 @@
-import { Molecula } from './components/rup/moleculas/molecula.component';
-import { Atomo } from './components/rup/atomos/atomoComponent';
-
 /*
 @jgabriel | 04-03-2017
 
@@ -10,9 +7,10 @@ de la siguiente manera:
 
 1) Módulos principales de Angular
 2) Módulos globales
-3) Servicios
-4) Componentes
-5) Otros
+3) Pipes
+4) Servicios
+5) Componentes
+6) Otros
 */
 
 // Angular
@@ -30,9 +28,24 @@ import { Auth } from '@andes/auth';
 import { RoutingGuard } from './app.routings-guard.class';
 import { AgmCoreModule } from 'angular2-google-maps/core';
 import { MapsComponent } from './utils/mapsComponent';
+import { PermisosComponent } from './utils/permisos/permisos.component';
+import { Ng2DragDropModule } from 'ng2-drag-drop';
+import { HoverClassDirective } from './directives/hover-class.directive';
+import { ChartsModule } from 'ng2-charts';
 
-//Pipes
-import { patientFullNamePipe, patientRealAgePipe } from './utils/patientPipe';
+// Pipes
+import { EdadPipe } from './pipes/edad.pipe';
+import { ProfesionalPipe } from './pipes/profesional.pipe';
+import { FromNowPipe } from './pipes/fromNow.pipe';
+import { FechaPipe } from './pipes/fecha.pipe';
+import { PacientePipe } from './pipes/paciente.pipe';
+import { SexoPipe } from './pipes/sexo.pipe';
+import { OrganizacionPipe } from './pipes/organizacion.pipe';
+import { SortBloquesPipe } from './pipes/agenda-bloques.pipe';
+import { TextFilterPipe } from './pipes/textFilter.pipe';
+import { FilterPermisos } from './pipes/filterPermisos.pipe';
+import { EnumerarPipe } from './pipes/enumerar.pipe';
+import { PluralizarPipe } from './pipes/pluralizar.pipe';
 
 // Servicios
 // ... Tablas Maestras
@@ -46,24 +59,48 @@ import { PacienteService } from './services/paciente.service';
 import { TipoEstablecimientoService } from './services/tipoEstablecimiento.service';
 import { ProvinciaService } from './services/provincia.service';
 import { FinanciadorService } from './services/financiador.service';
+import { ParentescoService } from './services/parentesco.service';
 import { ListaEsperaService } from './services/turnos/listaEspera.service';
+import { LogService } from './services/log.service';
+import { LogPacienteService } from './services/logPaciente.service';
+import { PermisosService } from './services/permisos.service';
 
 // ... Turnos
-import { TurnosComponent } from './components/turnos/turnos.component';
 import { EspacioFisicoService } from './services/turnos/espacio-fisico.service';
 import { AgendaService } from './services/turnos/agenda.service';
+import { AppMobileService } from './services/appMobile.service';
 import { TurnoService } from './services/turnos/turno.service';
 import { SmsService } from './services/turnos/sms.service';
 import { ConfigPrestacionService } from './services/turnos/configPrestacion.service';
 import { TipoPrestacionService } from './services/tipoPrestacion.service';
-// ... RUP
-import { TipoProblemaService } from './services/rup/tipoProblema.service';
-import { ProblemaPacienteService } from './services/rup/problemaPaciente.service';
-import { PrestacionPacienteService } from './services/rup/prestacionPaciente.service';
+
+// ... Usuarios
+import { UsuarioService } from './services/usuarios/usuario.service';
+
+// ... term
+import { Cie10Service } from './services/term/cie10.service';
+
+// SNOMED
+import { SnomedService } from './services/term/snomed.service';
+
+// ... Llaves
+import { LlavesTipoPrestacionService } from './services/llaves/llavesTipoPrestacion.service';
+
+// AUDITORIA
+import { AuditoriaPorBloqueService } from './services/auditoria/auditoriaPorBloque.service';
+import { AuditoriaService } from './services/auditoria/auditoria.service';
+
+// Auditoría
+import { AuditoriaPrestacionPacienteService } from './services/auditoria/auditoriaPrestacionPaciente.service';
+import { SisaService } from './services/fuentesAutenticas/servicioSisa.service';
+import { SintysService } from './services/fuentesAutenticas/servicioSintys.service';
+import { AnsesService } from './services/fuentesAutenticas/servicioAnses.service';
 
 // Componentes
-// ... Tablas Maestras
+import { LoginComponent } from './components/login/login.component';
+import { SelectOrganizacionComponent } from './components/login/selectOrganizacion.component';
 import { InicioComponent } from './components/inicio/inicio.component';
+// ... Tablas Maestras
 import { ProfesionalComponent } from './components/profesional/profesional.component';
 import { ProfesionalCreateUpdateComponent } from './components/profesional/profesional-create-update.component';
 import { EspecialidadComponent } from './components/especialidad/especialidad.component';
@@ -76,238 +113,297 @@ import { TipoPrestacionCreateUpdateComponent } from './components/tipoPrestacion
 import { PacienteSearchComponent } from './components/paciente/paciente-search.component';
 import { PacienteCreateUpdateComponent } from './components/paciente/paciente-create-update.component';
 import { HeaderPacienteComponent } from './components/paciente/headerPaciente.component';
+import { DashboardComponent } from './components/paciente/dashboard.component';
+import { PacienteDetalleComponent } from './components/paciente/paciente-detalle';
 
 // ... Turnos
-import { ClonarAgendaComponent } from './components/turnos/clonar-agenda';
-import { AgendaComponent } from './components/turnos/agenda.component';
+import { TurnosComponent } from './components/turnos/gestor-agendas/turnos.component';
+import { ClonarAgendaComponent } from './components/turnos/gestor-agendas/operaciones-agenda/clonar-agenda';
+import { PlanificarAgendaComponent } from './components/turnos/gestor-agendas/operaciones-agenda/planificar-agenda.component';
 import { DarTurnosComponent } from './components/turnos/dar-turnos/dar-turnos.component';
-import { GestorAgendasComponent } from './components/turnos/gestor-agendas.component';
 import { CalendarioComponent } from './components/turnos/dar-turnos/calendario.component';
-import { VistaAgendaComponent } from './components/turnos/vista-agenda.component';
 import { PanelEspacioComponent } from './components/turnos/configuracion/espacio-fisico/panel-espacio.component';
 import { ListaEsperaCreateUpdateComponent } from './components/turnos/lista-espera/listaEspera-create-update.component';
 import { ListaEsperaComponent } from './components/turnos/lista-espera/listaEspera.component';
-import { LiberarTurnoComponent } from './components/turnos/liberar-turno.component';
-import { SuspenderTurnoComponent } from './components/turnos/suspender-turno.component';
+import { LiberarTurnoComponent } from './components/turnos/gestor-agendas/operaciones-turnos/liberar-turno.component';
+import { SuspenderTurnoComponent } from './components/turnos/gestor-agendas/operaciones-turnos/suspender-turno.component';
+import { ReasignarTurnoComponent } from './components/turnos/gestor-agendas/operaciones-turnos/reasignar/reasignar-turno.component';
+import { ReasignarTurnoAutomaticoComponent } from './components/turnos/gestor-agendas/operaciones-turnos/reasignar/reasignar-turno-automatico.component';
+import { ReasignarTurnoAgendasComponent } from './components/turnos/gestor-agendas/operaciones-turnos/reasignar/reasignar-turno-agendas.component';
 import { EspacioFisicoComponent } from './components/turnos/configuracion/espacio-fisico/espacio-fisico.component';
 import { EditEspacioFisicoComponent } from './components/turnos/configuracion/espacio-fisico/edit-espacio-fisico.component';
-import { AgregarNotaTurnoComponent } from './components/turnos/agregar-nota-turno.component';
-import { PanelAgendaComponent } from './components/turnos/panel-agenda.component';
-
-// Estos componentes utilizan ng-prime y deben ser actualizados
-// import { PrestacionComponent } from './components/turnos/configuracion/prestacion/prestacion.component';
-// import { PrestacionCreateComponent } from './components/turnos/configuracion/prestacion/prestacion-create.component';
-// import { PrestacionUpdateComponent } from './components/turnos/configuracion/prestacion/prestacion-update.component';
+import { FiltrosMapaEspacioFisicoComponent } from './components/turnos/configuracion/mapa-espacio-fisico/filtros-mapa-espacio-fisico.component';
+import { AgregarNotaTurnoComponent } from './components/turnos/gestor-agendas/operaciones-turnos/agregar-nota-turno.component';
+import { CarpetaPacienteComponent } from './components/turnos/gestor-agendas/operaciones-turnos/carpeta-paciente.component';
+import { GestorAgendasComponent } from './components/turnos/gestor-agendas/gestor-agendas.component';
+import { AgregarNotaAgendaComponent } from './components/turnos/gestor-agendas/operaciones-agenda/nota-agenda.component';
+import { AgregarSobreturnoComponent } from './components/turnos/gestor-agendas/operaciones-agenda/sobreturno.component';
+import { PanelAgendaComponent } from './components/turnos/gestor-agendas/operaciones-agenda/panel-agenda.component';
+import { BotonesAgendaComponent } from './components/turnos/gestor-agendas/operaciones-agenda/botones-agenda.component';
+import { RevisionAgendaComponent } from './components/turnos/gestor-agendas/operaciones-agenda/revision-agenda.component';
+import { PopoverAuditComponent } from './components/popover-audit/popover-audit.component';
+import { PuntoInicioTurnosComponent } from './components/turnos/dashboard/puntoInicio-turnos.component';
+import { EstadisticasAgendasComponent } from './components/turnos/dashboard/estadisticas-agendas.component';
+import { EstadisticasPacientesComponent } from './components/turnos/dashboard/estadisticas-pacientes.component';
+import { PacienteSearchTurnosComponent } from './components/turnos/dashboard/paciente-search-turnos.component';
+import { TurnosPacienteComponent } from './components/turnos/dashboard/turnos-paciente.component';
+import { DashboardCodificacionComponent } from './components/turnos/dashboard/dashboard-codificacion.component';
+import { ActivarAppComponent } from './components/turnos/dashboard/activar-app.component';
+import { SolicitudTurnoVentanillaComponent } from './components/turnos/dashboard/solicitud-turno-ventanilla/solicitud-turno-ventanilla.component';
+import { ListaSolicitudTurnoVentanillaComponent } from './components/turnos/dashboard/solicitud-turno-ventanilla/lista-solicitud-turno-ventanilla.component';
+import { ListarTurnosComponent } from './components/turnos/gestor-agendas/operaciones-agenda/listar-turnos.component';
+import { MapaEspacioFisicoComponent } from './components/turnos/configuracion/mapa-espacio-fisico/mapa-espacio-fisico.component';
 
 // ... RUP
-import { RupComponent } from './components/rup/rup.component';
-import { PrestacionValidacionComponent } from './components/rup/ejecucion/prestacionValidacion.component';
-import { PesoComponent } from './components/rup/atomos/peso.component';
-import { SaturacionOxigenoComponent } from './components/rup/atomos/saturacionOxigeno.component';
-import { TallaComponent } from './components/rup/atomos/talla.component';
-import { TemperaturaComponent } from './components/rup/atomos/temperatura.component';
-import { FrecuenciaCardiacaComponent } from './components/rup/atomos/frecuenciaCardiaca.component';
-import { FrecuenciaRespiratoriaComponent } from './components/rup/atomos/frecuenciaRespiratoria.component';
-import { TensionSistolicaComponent } from './components/rup/atomos/tensionSistolica.component';
-import { TensionDiastolicaComponent } from './components/rup/atomos/tensionDiastolica.component';
-import { SignosVitalesComponent } from './components/rup/moleculas/signos-vitales/signosVitales.component';
-import { TensionArterialComponent } from './components/rup/moleculas/tension-arterial/tensionArterial.component';
-import { PrestacionEjecucionComponent } from './components/rup/ejecucion/prestacionEjecucion.component';
-import { ResumenComponent } from './components/rup/ejecucion/resumen.component';
-import { EvolucionTodosProblemasComponent } from './components/rup/problemas-paciente/evolucionTodosProblemas.component';
-import { EvolucionProblemaComponent } from './components/rup/problemas-paciente/evolucionProblema.component';
-import { EnmendarProblemaComponent } from './components/rup/problemas-paciente/enmendarProblema.component';
-import { TransformarProblemaComponent } from './components/rup/problemas-paciente/transformarProblema.component';
-import { verProblemaComponent } from './components/rup/problemas-paciente/verProblema.component';
-import { ConsultaGeneralClinicaMedicaComponent } from './components/rup/moleculas/consulta-general-clinica-medica/consultaGeneralClinicaMedica.component';
-import { ObservacionesComponent } from './components/rup/atomos/observaciones.component';
-import { PuntoInicioComponent } from './components/rup/ejecucion/puntoInicio.component';
-import { EdadGestacionalComponent } from './components/rup/atomos/perinatales-nacimiento/edadGestacional.component';
-import { ScoreApgarComponent } from './components/rup/atomos/perinatales-nacimiento/scoreApgar.component';
-import { NacimientoComponent } from './components/rup/moleculas/nacimiento/nacimiento.component';
-import { EdadGestacionalFetalComponent } from './components/rup/moleculas/edad-gestacional-fetal/edadGestacionalFetal.component';
-import { PesquisaNeonatalComponent } from './components/rup/atomos/pesquisaNeonatal.component';
-import { PruebaOtoemisionesAcusticasComponent } from './components/rup/atomos/pruebaOtoemisionesAcusticas.component';
-import { PercentiloTallaComponent } from './components/rup/atomos/percentiloTalla.component';
-import { PercentiloCircunferenciaCefalicaNinoComponent } from './components/rup/atomos/percentiloCircunferenciaCefalicaNino.component';
-import { DesarrolloMotorComponent } from './components/rup/atomos/desarrolloMotor.component';
-import { ActitudAnteLosCuidadosComponent } from './components/rup/atomos/actitudAnteLosCuidados.component';
-import { ControlDeEsfinteresComponent } from './components/rup/atomos/controlDeEsfinteres.component';
-import { DesarrolloIntelectualyJuegosComponent } from './components/rup/atomos/desarrolloIntelectualyJuegos.component';
-import { EscalaDeDesarrolloComponent } from './components/rup/moleculas/escala-de-desarrollo/escalaDeDesarrollo.component';
-import { FactoresDeRiesgoNinoSanoComponent } from './components/rup/atomos/factores-riesgo/factoresDeRiesgoNinoSano.component';
-import { IndiceDeMasaCorporalComponent } from './components/rup/moleculas/indice-de-masa-corporal/indiceDeMasaCorporal.component';
-
-// ATOMO SOCIOECONOMICO
-import { ViviendaSituacionSocioEconomicaComponent } from './components/rup/moleculas/vivienda-situacion-socioeconomica-familiar/viviendaSituacionSocioEconomica.component';
-import { ViviendaResiduosComponent } from './components/rup/atomos/socio-economicos/viviendaResiduos.component';
-import { ViviendaPisoComponent } from './components/rup/atomos/socio-economicos/viviendaPiso.component';
-import { ViviendaNivelInstruccionComponent } from './components/rup/atomos/socio-economicos/viviendaNivelInstruccion.component';
-import { ViviendaFamiliaComponent } from './components/rup/atomos/socio-economicos/viviendaFamilia.component';
-import { ViviendaContaminantesComponent } from './components/rup/atomos/socio-economicos/viviendaContaminantes.component';
-import { ViviendaCombustionComponent } from './components/rup/atomos/socio-economicos/viviendaCombustion.component';
-import { ViviendaSostenEconomicoComponent } from './components/rup/atomos/socio-economicos/viviendaSostenEconomico.component';
-import { ViviendaAsistenciaEconomicaComponent } from './components/rup/atomos/socio-economicos/viviendaAsistenciaEconomica.component';
-import { ViviendaCondicionesAlojamientoComponent } from './components/rup/moleculas/vivienda-condiciones-y-alojamiento/viviendaCondicionesAlojamiento.component';
-// PERINATALES EMBARAZO
-import { PerinatalesEmbarazoNormalComponent } from './components/rup/atomos/perinatales-embarazo/perinatalesEmbarazoNormal.component';
-import { PerinatalesEmbarazoAnormalComponent } from './components/rup/atomos/perinatales-embarazo/perinatalesEmbarazoAnormal.component';
-import { PerinatalesNumeroGestaComponent } from './components/rup/atomos/perinatales-embarazo/perinatalesNumeroGesta.component';
-import { PerinatalesGestacionMultipleComponent } from './components/rup/atomos/perinatales-embarazo/perinatalesGestacionMultiple.component';
-// PERINATALES PARTO
-import { PartoViaVaginalComponent } from './components/rup/atomos/perinatales-parto/partoViaVaginal.component';
-import { PartoViaVaginalForcepsComponent } from './components/rup/atomos/perinatales-parto/partoViaVaginalForceps.component';
-import { PartoVaginalAsistidoExtractorVacioComponent } from './components/rup/atomos/perinatales-parto/partoVaginalAsistidoExtractorVacio.component';
-import { PartoCesareaComponent } from './components/rup/atomos/perinatales-parto/partoCesarea.component';
-import { HallazgoRelacionadoPartoComponent } from './components/rup/atomos/perinatales-parto/hallazgoRelacionadoParto.component';
-// NIÑO SANO
-import { OdontologiaComponent } from './components/rup/atomos/odontologia.component';
-import { EstadoNutricionalComponent } from './components/rup/atomos/estadoNutricional.component';
+import { RUPComponent } from './modules/rup/components/core/rup.component';
+import { ElementosRUPService } from './modules/rup/services/elementosRUP.service';
+import { BuscadorComponent } from './modules/rup/components/ejecucion/buscador.component';
+import { HudsBusquedaComponent } from './modules/rup/components/ejecucion/hudsBusqueda.component';
+import { PrestacionesService } from './modules/rup/services/prestaciones.service';
+import { ConceptObserverService } from './modules/rup/services/conceptObserver.service';
+import { PrestacionCrearComponent } from './modules/rup/components/ejecucion/prestacionCrear.component';
+import { SnomedBuscarComponent } from './components/snomed/snomed-buscar.component';
+import { ResumenComponent } from './modules/rup/components/ejecucion/resumen.component';
+import { PrestacionValidacionComponent } from './modules/rup/components//ejecucion/prestacionValidacion.component';
+import { PrestacionEjecucionComponent } from './modules/rup/components//ejecucion/prestacionEjecucion.component';
+import { PuntoInicioComponent } from './modules/rup/components/ejecucion/puntoInicio.component';
+// import { RUPRegistry } from './modules/rup/components/core/rup-.registry';
+// TODO: ver con JGabriel!!!
+import { TensionSistolicaComponent } from './modules/rup/components/elementos/tensionSistolica.component';
+import { TensionDiastolicaComponent } from './modules/rup/components/elementos/tensionDiastolica.component';
+import { TensionArterialComponent } from './modules/rup/components/elementos/tensionArterial.component';
+import { TemperaturaComponent } from './modules/rup/components/elementos/temperatura.component';
+import { TallaComponent } from './modules/rup/components/elementos/talla.component';
+import { SolicitudPrestacionDefaultComponent } from './modules/rup/components/elementos/solicitudPrestacionDefault.component';
+import { SignosVitalesComponent } from './modules/rup/components/elementos/signosVitales.component';
+import { SaturacionOxigenoComponent } from './modules/rup/components/elementos/saturacionOxigeno.component';
+import { FrecuenciaCardiacaComponent } from './modules/rup/components/elementos/frecuenciaCardiaca.component';
+import { FrecuenciaRespiratoriaComponent } from './modules/rup/components/elementos/frecuenciaRespiratoria.component';
+import { PesoComponent } from './modules/rup/components/elementos/peso.component';
+import { ObservacionesComponent } from './modules/rup/components/elementos/observaciones.component';
+import { NuevaEvolucionProblemaComponent } from './modules/rup/components/elementos/nuevaEvolucionProblema.component';
+import { IndiceDeMasaCorporalComponent } from './modules/rup/components/elementos/indiceDeMasaCorporal.component';
+import { EvolucionProblemaDefaultComponent } from './modules/rup/components/elementos/evolucionProblemaDefault.component';
+import { AutocitadoComponent } from './modules/rup/components/elementos/autocitado.component';
 
 
+// Llaves
+import { LlavesTipoPrestacionComponent } from './components/llaves/tipoPrestacion/llaves-tipoPrestacion.component';
+import { EditarLlavesTipoPrestacionComponent } from './components/llaves/tipoPrestacion/editar-llaves-tipoPrestacion.component';
 
-export const RUP_COMPONENTS = [
-  TensionArterialComponent,
-  SignosVitalesComponent,
-  FrecuenciaCardiacaComponent,
-  FrecuenciaRespiratoriaComponent,
-  PesoComponent,
-  SaturacionOxigenoComponent,
-  TallaComponent,
-  TemperaturaComponent,
-  TensionSistolicaComponent,
-  TensionDiastolicaComponent,
-  ObservacionesComponent,
-  ConsultaGeneralClinicaMedicaComponent,
-  EvolucionProblemaComponent,
-  EvolucionTodosProblemasComponent,
-  PuntoInicioComponent,
-  PrestacionEjecucionComponent,
-  ResumenComponent,
-  HeaderPacienteComponent,
-  PrestacionValidacionComponent,
-  EdadGestacionalComponent,
-  ScoreApgarComponent,
-  NacimientoComponent,
-  EdadGestacionalFetalComponent,
-  ViviendaAsistenciaEconomicaComponent,
-  ViviendaCombustionComponent,
-  ViviendaContaminantesComponent,
-  ViviendaFamiliaComponent,
-  ViviendaPisoComponent,
-  ViviendaResiduosComponent,
-  ViviendaSituacionSocioEconomicaComponent,
-  ViviendaSostenEconomicoComponent,
-  ViviendaNivelInstruccionComponent,
-  ViviendaCondicionesAlojamientoComponent,
-  PerinatalesEmbarazoNormalComponent,
-  PerinatalesEmbarazoAnormalComponent,
-  PerinatalesNumeroGestaComponent,
-  PerinatalesGestacionMultipleComponent,
-  HallazgoRelacionadoPartoComponent,
-  PartoCesareaComponent,
-  PartoVaginalAsistidoExtractorVacioComponent,
-  PartoViaVaginalForcepsComponent,
-  PartoViaVaginalComponent,
-  EvolucionProblemaComponent,
-  EnmendarProblemaComponent,
-  TransformarProblemaComponent,
-  verProblemaComponent,
-  PruebaOtoemisionesAcusticasComponent,
-  PesquisaNeonatalComponent,
-  PercentiloTallaComponent,
-  PercentiloCircunferenciaCefalicaNinoComponent,
-  DesarrolloMotorComponent,
-  ActitudAnteLosCuidadosComponent,
-  ControlDeEsfinteresComponent,
-  DesarrolloIntelectualyJuegosComponent,
-  EscalaDeDesarrolloComponent,
-  EstadoNutricionalComponent,
-  OdontologiaComponent,
-  IndiceDeMasaCorporalComponent,
-  FactoresDeRiesgoNinoSanoComponent,
-  Atomo,
-  Molecula
-  
+// ... Auditoría RUP (prestacionPaciente)
+import { AuditoriaPrestacionPacienteComponent } from './components/auditoria/prestacionPaciente/auditoria-prestacionPaciente.component';
+import { EditarAuditoriaPrestacionPacienteComponent } from './components/auditoria/prestacionPaciente/editar-auditoria-prestacionPaciente.component';
 
+
+// AUDITORIA
+import { AuditoriaComponent } from './components/auditoria/auditoria.component';
+import { Auditoria2Component } from './components/auditoria/auditoria2.component';
+// import { AuditoriaPorBloqueComponent } from './components/auditoria/auditoriaPorBloque.component';
+
+// USUARIO
+import { BusquedaUsuarioComponent } from './components/usuario/busquedaUsuario.component';
+import { UsuarioCreateComponent } from './components/usuario/usuarioCreate.component';
+import { UsuarioUpdateComponent } from './components/usuario/usuarioUpdate.component';
+import { ArbolPermisosComponent } from './components/usuario/arbolPermisos.component';
+
+
+export const RUP_ELEMENTS: any[] = [
+    { component: BuscadorComponent, key: 'BuscadorComponent' },
+    { component: FrecuenciaCardiacaComponent, key: 'FrecuenciaCardiacaComponent' },
+    { component: FrecuenciaRespiratoriaComponent, key: 'FrecuenciaRespiratoriaComponent' },
+    { component: HeaderPacienteComponent, key: 'HeaderPacienteComponent' },
+    { component: HudsBusquedaComponent, key: 'HudsBusquedaComponent' },
+    { component: IndiceDeMasaCorporalComponent, key: 'IndiceDeMasaCorporalComponent' },
+    { component: NuevaEvolucionProblemaComponent, key: 'NuevaEvolucionProblemaComponent' },
+    { component: ObservacionesComponent, key: 'ObservacionesComponent' },
+    { component: PesoComponent, key: 'PesoComponent' },
+    { component: PrestacionCrearComponent, key: 'PrestacionCrearComponent' },
+    { component: PrestacionEjecucionComponent, key: 'PrestacionEjecucionComponent' },
+    { component: PrestacionValidacionComponent, key: 'PrestacionValidacionComponent' },
+    { component: PuntoInicioComponent, key: 'PuntoInicioComponent' },
+    { component: ResumenComponent, key: 'ResumenComponent' },
+    { component: SaturacionOxigenoComponent, key: 'SaturacionOxigenoComponent' },
+    { component: SignosVitalesComponent, key: 'SignosVitalesComponent' },
+    { component: SnomedBuscarComponent, key: 'SnomedBuscarComponent' },
+    { component: SolicitudPrestacionDefaultComponent, key: 'SolicitudPrestacionDefaultComponent' },
+    { component: TallaComponent, key: 'TallaComponent' },
+    { component: TemperaturaComponent, key: 'TemperaturaComponent' },
+    { component: TensionArterialComponent, key: 'TensionArterialComponent' },
+    { component: TensionDiastolicaComponent, key: 'TensionDiastolicaComponent' },
+    { component: TensionSistolicaComponent, key: 'TensionSistolicaComponent' },
+    { component: AutocitadoComponent, key: 'AutocitadoComponent' },
+
+];
+
+// let temp = RUP_ELEMENTS.map((i) => i.component);
+const RUP_COMPONENTS = [
+    BuscadorComponent,
+    EvolucionProblemaDefaultComponent,
+    FrecuenciaCardiacaComponent,
+    FrecuenciaRespiratoriaComponent,
+    HeaderPacienteComponent,
+    HudsBusquedaComponent,
+    IndiceDeMasaCorporalComponent,
+    NuevaEvolucionProblemaComponent,
+    ObservacionesComponent,
+    PesoComponent,
+    PrestacionCrearComponent,
+    PrestacionEjecucionComponent,
+    PrestacionValidacionComponent,
+    PuntoInicioComponent,
+    ResumenComponent,
+    SaturacionOxigenoComponent,
+    SignosVitalesComponent,
+    SnomedBuscarComponent,
+    SolicitudPrestacionDefaultComponent,
+    TallaComponent,
+    TemperaturaComponent,
+    TensionArterialComponent,
+    TensionDiastolicaComponent,
+    TensionSistolicaComponent,
+    AutocitadoComponent
 ];
 
 // Locales
 import { AppComponent } from './app.component';
 import { routing, appRoutingProviders } from './app.routing';
 
-// Ver donde poner
-import { ConfirmationPopoverModule } from 'angular-confirmation-popover';
+// Componentes RUP
+// [jgabriel] Por alguna cuestión de Angular's DI no se puede tener esto en otro archivo. WTF?
 
+export let RUPRegistry = {
+    'PesoComponent': PesoComponent,
+    'EvolucionProblemaDefaultComponent': EvolucionProblemaDefaultComponent,
+    'IndiceDeMasaCorporalComponent': IndiceDeMasaCorporalComponent,
+    'NuevaEvolucionProblemaComponent': NuevaEvolucionProblemaComponent,
+    'ObservacionesComponent': ObservacionesComponent,
+    'SaturacionOxigenoComponent': SaturacionOxigenoComponent,
+    'SignosVitalesComponent': SignosVitalesComponent,
+    'SolicitudPrestacionDefaultComponent': SolicitudPrestacionDefaultComponent,
+    'TallaComponent': TallaComponent,
+    'TemperaturaComponent': TemperaturaComponent,
+    'TensionArterialComponent': TensionArterialComponent,
+    'TensionDiastolicaComponent': TensionDiastolicaComponent,
+    'TensionSistolicaComponent': TensionSistolicaComponent,
+    'FrecuenciaCardiacaComponent': FrecuenciaCardiacaComponent,
+    'FrecuenciaRespiratoriaComponent': FrecuenciaRespiratoriaComponent,
+    'AutocitadoComponent': AutocitadoComponent
+
+};
+
+let RUPComponentsArray = [];
+for (let key in RUPRegistry) {
+    RUPComponentsArray.push(RUPRegistry[key]);
+}
 
 // Main module
 @NgModule({
-  imports: [
-    BrowserModule,
-    ReactiveFormsModule,
-    FormsModule,
-    HttpModule,
-    PlexModule,
-    routing,
-    AgmCoreModule.forRoot({
-      apiKey: 'AIzaSyAJuFVuMmVwV8gtP_1m3Ll1VzHagAI_X9I'
-    }),
-    ConfirmationPopoverModule.forRoot({
-      confirmButtonType: 'danger' // set defaults here
-    })
-  ],
+    imports: [
+        BrowserModule,
+        ReactiveFormsModule,
+        FormsModule,
+        HttpModule,
+        PlexModule,
+        Ng2DragDropModule,
+        ChartsModule,
+        routing,
+        AgmCoreModule.forRoot({
+            apiKey: 'AIzaSyAJuFVuMmVwV8gtP_1m3Ll1VzHagAI_X9I'
+        })
+    ],
+    declarations: [
+        AppComponent, InicioComponent, LoginComponent, SelectOrganizacionComponent,
+        OrganizacionComponent, OrganizacionCreateUpdateComponent,
+        ProfesionalComponent, ProfesionalCreateUpdateComponent,
+        ProfesionalCreateUpdateComponent,
+        EspecialidadComponent, EspecialidadCreateUpdateComponent,
+        PacienteCreateUpdateComponent, PacienteDetalleComponent, PacienteSearchComponent, DashboardComponent,
+        MapsComponent, EdadPipe, ProfesionalPipe, FromNowPipe, FechaPipe, PacientePipe, SexoPipe, OrganizacionPipe, SortBloquesPipe, TextFilterPipe,
+        FilterPermisos, EnumerarPipe, PluralizarPipe,
+        PlanificarAgendaComponent, PanelEspacioComponent, EspacioFisicoComponent, EditEspacioFisicoComponent, FiltrosMapaEspacioFisicoComponent,
+        TipoPrestacionComponent, TipoPrestacionCreateUpdateComponent,
+        DarTurnosComponent, CalendarioComponent, GestorAgendasComponent,
+        TurnosComponent, BotonesAgendaComponent, ClonarAgendaComponent,
+        ListaEsperaComponent, ListaEsperaCreateUpdateComponent, RevisionAgendaComponent, PopoverAuditComponent,
+        RUPComponent, LiberarTurnoComponent, SuspenderTurnoComponent, AgregarNotaTurnoComponent, AgregarNotaAgendaComponent,
+        AgregarSobreturnoComponent, PanelAgendaComponent,
+        CarpetaPacienteComponent,
+        ReasignarTurnoComponent, ReasignarTurnoAutomaticoComponent, EstadisticasAgendasComponent, EstadisticasPacientesComponent,
+        AuditoriaComponent,
+        PermisosComponent, ArbolPermisosComponent,
+        // AuditoriaPorBloqueComponent,
+        PuntoInicioComponent,
+        Auditoria2Component,
+        LlavesTipoPrestacionComponent, EditarLlavesTipoPrestacionComponent,
+        AuditoriaPrestacionPacienteComponent, EditarAuditoriaPrestacionPacienteComponent,
+        HoverClassDirective, PuntoInicioTurnosComponent, ReasignarTurnoAgendasComponent,
+        PacienteSearchTurnosComponent, TurnosPacienteComponent, DashboardCodificacionComponent,
+        SolicitudTurnoVentanillaComponent, ListaSolicitudTurnoVentanillaComponent, ActivarAppComponent,
+        BusquedaUsuarioComponent, UsuarioCreateComponent, UsuarioUpdateComponent,
+        ListarTurnosComponent,
+        MapaEspacioFisicoComponent,
+        ResumenComponent,
+        PrestacionCrearComponent,
+        PrestacionEjecucionComponent,
+        PrestacionValidacionComponent,
+        SnomedBuscarComponent,
+        HeaderPacienteComponent,
+        HudsBusquedaComponent,
+        BuscadorComponent,
+        // RUP
+        ...RUPComponentsArray,
+    ],
+    entryComponents: RUPComponentsArray,
+    bootstrap: [AppComponent],
+    providers: [
+        {
+            provide: LOCALE_ID,
+            useValue: 'es-AR'
+        },
+        Plex,
+        Auth,
+        RoutingGuard,
+        OrganizacionService,
+        ProvinciaService,
+        TipoEstablecimientoService,
+        EspecialidadService,
+        ProfesionalService,
+        PaisService,
+        LocalidadService,
+        BarrioService,
+        PacienteService,
+        FinanciadorService,
+        ParentescoService,
+        appRoutingProviders,
+        ConfigPrestacionService,
+        PlanificarAgendaComponent,
+        AgendaService,
+        AppMobileService,
+        TurnoService,
+        EspacioFisicoService,
+        ListaEsperaService,
+        Server,
+        SmsService,
+        PrestacionesService,
+        TipoPrestacionService,
+        ElementosRUPService,
+        ConceptObserverService,
+        LlavesTipoPrestacionService,
+        LogService,
+        AuditoriaPorBloqueService,
+        AuditoriaService,
+        AuditoriaPrestacionPacienteService,
+        SnomedService,
+        Cie10Service,
+        SisaService,
+        SintysService,
+        AnsesService,
+        LogPacienteService,
+        UsuarioService,
+        PermisosService
 
-  declarations: [
-    AppComponent, InicioComponent,
-    OrganizacionComponent, OrganizacionCreateUpdateComponent,
-    ProfesionalComponent, ProfesionalCreateUpdateComponent,
-    ProfesionalCreateUpdateComponent,
-    EspecialidadComponent, EspecialidadCreateUpdateComponent,
-    PacienteCreateUpdateComponent, PacienteSearchComponent, MapsComponent, patientFullNamePipe, patientRealAgePipe,
-    AgendaComponent, PanelEspacioComponent, EspacioFisicoComponent, EditEspacioFisicoComponent,
-    // PrestacionComponent, PrestacionCreateComponent, PrestacionUpdateComponent,
-    TipoPrestacionComponent, TipoPrestacionCreateUpdateComponent,
-    DarTurnosComponent, CalendarioComponent, GestorAgendasComponent,
-    TurnosComponent, VistaAgendaComponent, ClonarAgendaComponent,
-    ListaEsperaComponent, ListaEsperaCreateUpdateComponent,
-    RupComponent, LiberarTurnoComponent, SuspenderTurnoComponent, AgregarNotaTurnoComponent,
-    PanelAgendaComponent,
-    ...RUP_COMPONENTS
-  ],
-  entryComponents: RUP_COMPONENTS,
-  bootstrap: [AppComponent],
-  providers: [{
-    provide: LOCALE_ID,
-    useValue: 'es-AR'
-  },
-    Plex,
-    Auth,
-    RoutingGuard,
-    OrganizacionService,
-    ProvinciaService,
-    TipoEstablecimientoService,
-    EspecialidadService,
-    ProfesionalService,
-    PaisService,
-    LocalidadService,
-    BarrioService,
-    PacienteService,
-    FinanciadorService,
-    appRoutingProviders,
-    ConfigPrestacionService,
-    AgendaComponent,
-    // EspacioFisicoComponent,
-    AgendaService,
-    TurnoService,
-    EspacioFisicoService,
-    ListaEsperaService,
-    Server,
-    SmsService,
-    PrestacionPacienteService,
-    ProblemaPacienteService,
-    TipoProblemaService,
-    TipoPrestacionService,
-  ]
+    ]
 })
 
 export class AppModule { }
