@@ -1,268 +1,266 @@
-import { BarrioService } from './../../services/barrio.service';
-import { IPais } from './../../interfaces/IPais';
-import { IBarrio } from './../../interfaces/IBarrio';
-import { ILocalidad } from './../../interfaces/ILocalidad';
+import { Plex } from '@andes/plex';
+import { Server } from '@andes/shared';
 import { Observable } from 'rxjs/Rx';
-import { OrganizacionService } from './../../services/organizacion.service';
-import { IOrganizacion } from './../../interfaces/IOrganizacion';
-import { Component, OnInit, Output, EventEmitter, Input } from '@angular/core';
-import { FormBuilder, FormGroup, Validators, FormArray } from '@angular/forms';
-import { ITipoEstablecimiento } from './../../interfaces/ITipoEstablecimiento';
-import { IProvincia } from './../../interfaces/IProvincia';
-import { TipoEstablecimientoService } from './../../services/tipoEstablecimiento.service';
+import { Component, OnInit, Output, EventEmitter, Input, HostBinding } from '@angular/core';
 import * as enumerados from './../../utils/enumerados';
+// Services
+import { BarrioService } from './../../services/barrio.service';
+import { TipoEstablecimientoService } from './../../services/tipoEstablecimiento.service';
+import { OrganizacionService } from './../../services/organizacion.service';
 import { PaisService } from './../../services/pais.service';
 import { ProvinciaService } from './../../services/provincia.service';
 import { LocalidadService } from './../../services/localidad.service';
+// Interfaces
+import { IPais } from './../../interfaces/IPais';
+import { IBarrio } from './../../interfaces/IBarrio';
+import { ILocalidad } from './../../interfaces/ILocalidad';
+import { IUbicacion } from './../../interfaces/IUbicacion';
+import { IEdificio } from './../../interfaces/IEdificio';
+import { IDireccion } from './../../interfaces/IDireccion';
+import { IContacto } from './../../interfaces/IContacto';
+import { IOrganizacion } from './../../interfaces/IOrganizacion';
+import { ITipoEstablecimiento } from './../../interfaces/ITipoEstablecimiento';
+import { IProvincia } from './../../interfaces/IProvincia';
 
 @Component({
     selector: 'organizacion-create-update',
     templateUrl: 'organizacion-create-update.html'
 })
 export class OrganizacionCreateUpdateComponent implements OnInit {
+
+    @HostBinding('class.plex-layout') layout = true;  // Permite el uso de flex-box en el componente
     @Input('seleccion') seleccion: IOrganizacion;
     @Output() data: EventEmitter<IOrganizacion> = new EventEmitter<IOrganizacion>();
 
-    createForm: FormGroup;
     // definición de arreglos
-    tipos: ITipoEstablecimiento[];
-    tiposcom: String[];
+    tiposEstablecimiento: ITipoEstablecimiento[];
     tipoComunicacion: any[];
-    paises: IPais[];
-    provincias: IProvincia[];
-    todasProvincias: IProvincia[];
-    localidades: ILocalidad[];
     todasLocalidades: ILocalidad[];
-    barrios: IBarrio[];
+    localidadesNeuquen: any[];
+
+    private paisArgentina = null;
+    private provinciaNeuquen = null;
+    private barrioNulleado = null;
+
+    tipoEstablecimiento: ITipoEstablecimiento = {
+        nombre: '',
+        descripcion: '',
+        clasificacion: '',
+        idTipoEfector: 0,
+    };
+
+    ubicacion: IUbicacion = {
+        barrio: {
+            id: '',
+            nombre: ''
+        },
+        localidad: {
+            id: '',
+            nombre: ''
+        },
+        provincia: {
+            id: '',
+            nombre: ''
+        },
+        pais: {
+            id: '',
+            nombre: ''
+        }
+    };
+
+    contacto: IContacto = {
+        tipo: 'celular',
+        valor: '',
+        ranking: 0,
+        activo: true,
+        ultimaActualizacion: new Date()
+    };
+
+    direccion: IDireccion = {
+        valor: '',
+        codigoPostal: '',
+        ubicacion: this.ubicacion,
+        ranking: 0,
+        geoReferencia: null,
+        ultimaActualizacion: new Date(),
+        activo: true
+    };
+
+    edificio: IEdificio = {
+        id: null,
+        descripcion: '',
+        contacto: this.contacto,
+        direccion: this.direccion,
+    };
+
+    organizacionModel: IOrganizacion = {
+        id: null,
+        codigo: {
+            sisa: '',
+            cuie: '',
+            remediar: ''
+        },
+        nombre: '',
+        tipoEstablecimiento: this.tipoEstablecimiento,
+        direccion: this.direccion,
+        contacto: [this.contacto],
+        edificio: [this.edificio],
+        nivelComplejidad: 0,
+        activo: true,
+        fechaAlta: new Date(),
+        fechaBaja: new Date(),
+    };
 
     constructor(
-        private formBuilder: FormBuilder,
         private organizacionService: OrganizacionService,
         private paisService: PaisService,
         private provinciaService: ProvinciaService,
         private localidadService: LocalidadService,
         private BarrioService: BarrioService,
-        private tipoEstablecimientoService: TipoEstablecimientoService
+        private tipoEstablecimientoService: TipoEstablecimientoService,
+        public plex: Plex, private server: Server
     ) { }
 
     ngOnInit() {
-        this.tiposcom = enumerados.getTipoComunicacion();
         this.tipoComunicacion = enumerados.getObjTipoComunicacion();
-        let nombre = this.seleccion ? this.seleccion.nombre : '';
-        let nivelComplejidad = this.seleccion ? this.seleccion.nivelComplejidad : '';
-        let sisa = this.seleccion ? this.seleccion.codigo.sisa : '';
-        let cuie = this.seleccion ? this.seleccion.codigo.cuie : '';
-        let remediar = this.seleccion ? this.seleccion.codigo.remediar : '';
-        let tipoEstablecimiento = this.seleccion ? this.seleccion.tipoEstablecimiento : '';
-        let valor = this.seleccion ? this.seleccion.direccion.valor : '';
-        let pais = this.seleccion ? this.seleccion.direccion.ubicacion.pais : '';
-        let provincia = this.seleccion ? this.seleccion.direccion.ubicacion.provincia : '';
-        let localidad = this.seleccion ? this.seleccion.direccion.ubicacion.localidad : '';
-        let codigoPostal = this.seleccion ? this.seleccion.direccion.codigoPostal : '';
-
-        this.createForm = this.formBuilder.group({
-            nombre: [nombre, Validators.required],
-            nivelComplejidad: [nivelComplejidad],
-            codigo: this.formBuilder.group({
-                sisa: [sisa, Validators.required],
-                cuie: [cuie],
-                remediar: [remediar],
-            }),
-            tipoEstablecimiento: [tipoEstablecimiento],
-            direccion: this.formBuilder.group({
-                valor: [valor],
-                ubicacion: this.formBuilder.group({
-                    pais: [pais],
-                    provincia: [provincia],
-                    localidad: [localidad]
-                }),
-                ranking: [''],
-                codigoPostal: [codigoPostal],
-                latitud: [''],
-                longitud: [''],
-                activo: [true]
-            }),
-            // telecom: this.formBuilder.array([]),
-            contacto: this.formBuilder.array([]),
-            edificio: this.formBuilder.array([])
+        this.tipoEstablecimientoService.get().subscribe(resultado => {
+            this.tiposEstablecimiento = resultado;
         });
 
-        if (this.seleccion) {
-            this.loadContactos();
-            /*this.seleccion.telecom.forEach(element => {
-                this.addTelecom(element);
-            });*/
-            this.loadEdificios();
-        }
-    }
-
-    /*Código de contactos*/
-
-    initContacto(rank: Number) {
-        // Inicializa contacto
-        let fecha = new Date();
-        return this.formBuilder.group({
-            tipo: ['', Validators.required],
-            valor: ['', Validators.required],
-            ranking: [rank],
-            ultimaActualizacion: [fecha],
-            activo: [true]
-        });
-    }
-
-    addContacto() {
-        const control = <FormArray>this.createForm.controls['contacto'];
-        control.push(this.initContacto(control.length + 1));
-    }
-
-    removeContacto(indice: number) {
-        const control = <FormArray>this.createForm.controls['contacto'];
-        control.removeAt(indice);
-    }
-
-    setContacto(cont: any) {
-        let tipo = cont ? enumerados.getObjeto(cont.tipo) : null;
-        return this.formBuilder.group({
-            tipo: [tipo, Validators.required],
-            valor: [cont.valor, Validators.required],
-            ranking: [cont.ranking],
-            ultimaActualizacion: [cont.ultimaActualizacion],
-            activo: [cont.activo]
-        });
-    }
-
-    loadContactos() {
-        let cantidadContactosActuales = this.seleccion.contacto.length;
-        const control = <FormArray>this.createForm.controls['contacto'];
-
-        if (cantidadContactosActuales > 0) {
-            for (let i = 0; i < cantidadContactosActuales; i++) {
-                let contacto: any = this.seleccion.contacto[i];
-                control.push(this.setContacto(contacto));
-            }
-        } else {
-            control.push(this.initContacto(1));
-        }
-    }
-
-    /*Cod. edificio*/
-
-    loadEdificios() {
-        this.seleccion.edificio.forEach(element => {
-            this.addEdificio(element, 'previo');
-        });
-    }
-    addEdificio(unEdificio, tipo) {
-        // agrega formContacto 
-        const control = <FormArray>this.createForm.controls['edificio'];
-        control.push(this.iniEdificio(unEdificio, 'nuevo'));
-    }
-
-    iniEdificio(unEdificio, tipo) {
-        // Inicializa edificio
-        if (tipo !== 'nuevo') {
-            return this.formBuilder.group({
-                id: unEdificio.id,
-                _id: unEdificio.id,
-                descripcion: unEdificio.descripcion,
-                direccion: this.formBuilder.group({
-                    valor: unEdificio.direccion ? unEdificio.direccion.valor : [''],
-                    ubicacion: this.formBuilder.group({
-                        pais: unEdificio.direccion ? unEdificio.direccion.ubicacion.pais :
-                            this.seleccion ? this.seleccion.direccion.ubicacion.pais : [''],
-                        provincia: unEdificio.direccion ? unEdificio.direccion.ubicacion.provincia :
-                            this.seleccion ? this.seleccion.direccion.ubicacion.provincia : [''],
-                        localidad: unEdificio.direccion ? unEdificio.direccion.ubicacion.localidad :
-                            this.seleccion ? this.seleccion.direccion.ubicacion.localidad : ['']
-                    }),
-                    ranking: unEdificio.direccion ? unEdificio.direccion.ranking : [''],
-                    codigoPostal: unEdificio.direccion ? unEdificio.direccion.codigoPostal :
-                        this.seleccion ? this.seleccion.direccion.codigoPostal : [''],
-                    latitud: [''],
-                    longitud: [''],
-                    activo: [true]
-                }),
-                telefono: this.formBuilder.group({
-                    tipo: unEdificio.telefono ? unEdificio.telefono.tipo : [''],
-                    valor: unEdificio.telefono ? unEdificio.telefono.valor : [''],
-                    ranking: unEdificio.telefono ? unEdificio.telefono.ranking : [''],
-                    activo: unEdificio.telefono ? unEdificio.telefono.activo : [''],
-                }),
-            });
-        } else {
-            return this.formBuilder.group({
-                descripcion: unEdificio.descripcion,
-                direccion: this.formBuilder.group({
-                    valor: unEdificio.direccion ? unEdificio.direccion.valor : [''],
-                    ubicacion: this.formBuilder.group({
-
-                        pais: unEdificio.direccion ? unEdificio.direccion.ubicacion.pais : this.seleccion ?
-                            this.seleccion.direccion.ubicacion.pais : [''],
-                        provincia: unEdificio.direccion ? unEdificio.direccion.ubicacion.provincia : this.seleccion ?
-                            this.seleccion.direccion.ubicacion.provincia : [''],
-                        localidad: unEdificio.direccion ? unEdificio.direccion.ubicacion.localidad : this.seleccion ?
-                            this.seleccion.direccion.ubicacion.localidad : ['']
-                    }),
-                    ranking: unEdificio.direccion ? unEdificio.direccion.ranking : [''],
-                    codigoPostal: unEdificio.direccion ? unEdificio.direccion.codigoPostal : this.seleccion ?
-                        this.seleccion.direccion.codigoPostal : [''],
-                    latitud: [''],
-                    longitud: [''],
-                    activo: [true]
-                }),
-                telefono: this.formBuilder.group({
-                    tipo: unEdificio.telefono ? unEdificio.telefono.tipo : [''],
-                    valor: unEdificio.telefono ? unEdificio.telefono.valor : [''],
-                    ranking: unEdificio.telefono ? unEdificio.telefono.ranking : [''],
-                    activo: unEdificio.telefono ? unEdificio.telefono.activo : [''],
-                }),
+        if (this.seleccion && this.seleccion.id) {
+            this.organizacionService.getById(this.seleccion.id).subscribe(resultado => {
+                if (resultado) {
+                    Object.assign(this.organizacionModel, resultado);
+                }
             });
         }
 
+        // Set País Argentina
+        this.paisService.get({
+            nombre: 'Argentina'
+        }).subscribe(arg => {
+            this.paisArgentina = arg[0];
+        });
+        // Set provincia Neuquen
+        this.provinciaService.get({
+            nombre: 'Neuquén'
+        }).subscribe(Prov => {
+            this.provinciaNeuquen = Prov[0];
+            this.loadLocalidades(this.provinciaNeuquen);
+        });
+
     }
 
-    removeEdificio(i: number) {
-        // elimina formContacto
-        const control = <FormArray>this.createForm.controls['edificio'];
-        control.removeAt(i);
-    }
+    onSave(valid) {
+        let organizacionGuardar = Object.assign({}, this.organizacionModel);
+        organizacionGuardar.contacto.map(elem => {
+            elem.tipo = ((typeof elem.tipo === 'string') ? elem.tipo : (Object(elem.tipo).id));
+            return elem;
+        });
+        if (organizacionGuardar.edificio) {
+            organizacionGuardar.edificio.forEach(e => {
+                if (e.contacto) {
+                    e.contacto.tipo = ((typeof e.contacto.tipo === 'string') ? e.contacto.tipo : (Object(e.contacto.tipo).id));
+                }
+                e.direccion.ubicacion.pais = this.paisArgentina;
+                e.direccion.ubicacion.provincia = this.provinciaNeuquen;
+                e.direccion.ubicacion.barrio = null;
+            });
+        }
+        organizacionGuardar.direccion.ubicacion.pais = this.paisArgentina;
+        organizacionGuardar.direccion.ubicacion.provincia = this.provinciaNeuquen;
+        organizacionGuardar.direccion.ubicacion.barrio = null;
 
-    loadTipos(event) {
-        this.tipoEstablecimientoService.get().subscribe(event.callback);
-    }
-
-    /*Código de filtrado de combos*/
-    loadPaises(event) {
-        this.paisService.get({}).subscribe(event.callback);
-    }
-
-    loadProvincias(event, pais) {
-        this.provinciaService.get({ pais: pais.value.id }).subscribe(event.callback);
-    }
-
-    loadLocalidades(event, provincia) {
-        this.localidadService.get({ provincia: provincia.value.id }).subscribe(event.callback);
-    }
-
-    onSave(model: any, isvalid: boolean) {
-        if (isvalid) {
-            let guardar: Observable<IOrganizacion>;
-            model.activo = true;
-            model.contacto = model.contacto.map(elem => { elem.tipo = elem.tipo.id; return elem; })
-            model.edificio = model.edificio.map(elem => { return elem; })
-            if (this.seleccion) {
-                model.id = this.seleccion.id;
-                guardar = this.organizacionService.put(model);
+        let operacion = this.organizacionService.save(organizacionGuardar);
+        operacion.subscribe(result => {
+            if (result) {
+                this.plex.alert('Los datos se actualizaron correctamente');
+                this.data.emit(result);
             } else {
-                guardar = this.organizacionService.post(model);
+                this.plex.alert('ERROR: Ocurrio un problema al actualizar los datos');
             }
-            guardar.subscribe(resultado => { this.data.emit(resultado); });
-        } else {
-            alert('Complete datos obligatorios');
-        }
+        });
     }
 
     onCancel() {
         window.setTimeout(() => this.data.emit(null), 100);
+    }
+
+    addContacto() {
+        let nuevoContacto = Object.assign({}, {
+            tipo: 'celular',
+            valor: '',
+            ranking: 0,
+            activo: true,
+            ultimaActualizacion: new Date()
+        });
+        this.organizacionModel.contacto.push(nuevoContacto);
+    }
+
+    removeContacto(i) {
+        if (i >= 0) {
+            this.organizacionModel.contacto.splice(i, 1);
+        }
+    }
+
+    addEdificio() {
+        let nuevoEdificio = Object.assign({}, {
+            id: null,
+            descripcion: '',
+            contacto: {
+                tipo: 'celular',
+                valor: '',
+                ranking: 0,
+                activo: true,
+                ultimaActualizacion: new Date()
+            },
+            direccion: {
+                valor: '',
+                codigoPostal: '',
+                ubicacion: {
+                    barrio: {
+                        id: '',
+                        nombre: ''
+                    },
+                    localidad: {
+                        id: '',
+                        nombre: ''
+                    },
+                    provincia: {
+                        id: '',
+                        nombre: ''
+                    },
+                    pais: {
+                        id: '',
+                        nombre: ''
+                    }
+                },
+                ranking: 0,
+                geoReferencia: null,
+                ultimaActualizacion: new Date(),
+                activo: true
+            },
+        });
+        if (this.organizacionModel.edificio) {
+            this.organizacionModel.edificio.push(nuevoEdificio);
+        } else {
+            this.organizacionModel.edificio = [nuevoEdificio];
+        }
+    }
+
+    removeEdificio(i) {
+        if (i >= 0) {
+            this.organizacionModel.edificio.splice(i, 1);
+        }
+    }
+
+    loadLocalidades(provincia) {
+        if (provincia && provincia.id) {
+            this.localidadService.get({
+                'provincia': provincia.id
+            }).subscribe(result => {
+                this.localidadesNeuquen = [...result];
+            });
+        }
     }
 }
