@@ -1,6 +1,6 @@
 import { IOrganizacion } from './../../interfaces/IOrganizacion';
 import { OrganizacionService } from './../../services/organizacion.service';
-import { Component, OnInit } from '@angular/core';
+import { Component, OnInit, HostBinding } from '@angular/core';
 import { FormBuilder, FormGroup } from '@angular/forms';
 
 const limit = 25;
@@ -10,18 +10,20 @@ const limit = 25;
     templateUrl: 'organizacion.html'
 })
 export class OrganizacionComponent implements OnInit {
-    showcreate: boolean = false;
+
+    @HostBinding('class.plex-layout') layout = true;  // Permite el uso de flex-box en el componente
+    showcreate = false;
     datos: IOrganizacion[] = [];
     searchForm: FormGroup;
     seleccion: IOrganizacion;
     value: any;
-    skip: number = 0;
-    nombre: string = '';
-    activo: Boolean = null;
-    loader: boolean = false;
-    finScroll: boolean = false;
-    tengoDatos: boolean = true;
-    checked: boolean = true;
+    skip = 0;
+    nombre = '';
+    activo: boolean = null;
+    loader = false;
+    finScroll = false;
+    tengoDatos = true;
+    checked = true;
 
     constructor(private formBuilder: FormBuilder, private organizacionService: OrganizacionService) { }
 
@@ -42,7 +44,8 @@ export class OrganizacionComponent implements OnInit {
     loadDatos(concatenar: boolean = false) {
         let parametros = {
             'activo': this.value && this.value.activo, 'nombre':
-            this.value && this.value.nombre, 'skip': this.skip, 'limit': limit
+            this.value && this.value.nombre
+            , 'skip': this.skip, 'limit': limit
         };
         this.organizacionService.get(parametros)
             .subscribe(
