@@ -14,17 +14,17 @@ const limit = 25;
 
 
 export class TipoPrestacionComponent implements OnInit {
-    showcreate: boolean = false;
-    showupdate: boolean = false;
+    showcreate = false;
+    showupdate = false;
     datos: ITipoPrestacion[];
     searchForm: FormGroup;
     seleccion: ITipoPrestacion;
-    skip: number = 0;
-    loader: boolean = false;
-    finScroll: boolean = false;
+    skip  = 0;
+    loader = false;
+    finScroll = false;
     value: any;
-    tengoDatos: boolean = true;
-    nombre: String= '';
+    tengoDatos = true;
+    nombre: String = '';
     granularidad:  {
                     id: '',
                     nombre: ''
@@ -84,14 +84,14 @@ export class TipoPrestacionComponent implements OnInit {
 
     activate(objTipoPrestacion: ITipoPrestacion) {
 
-        if (objTipoPrestacion.activo) {
+        /* if (objTipoPrestacion.activo) {
             this.tipoPrestacionService.disable(objTipoPrestacion)
                 .subscribe(datos => this.loadDatos());  // Bind to view
         } else {
             this.tipoPrestacionService.enable(objTipoPrestacion)
                 .subscribe(datos => this.loadDatos());  // Bind to view
-        }
+        } */
     }// activate
 
 
-}// export class TipoPrestacionComponent
+} // export class TipoPrestacionComponent
