@@ -6,17 +6,37 @@ import * as moment from 'moment';
 @Component({
     selector: 'app-calendario',
     templateUrl: 'calendario.html',
+    styleUrls: ['calendario.scss']
 })
 export class CalendarioComponent {
+    mostrarFinesDeSemana: any;
     private _agenda: any;
     private _agendas: Array<any>;
     private _estado: String;
-    private calendario: any = [];
     private diaSeleccionado: CalendarioDia;
+    public calendario: any = [];
+
+    private _opcionesCalendario;
+    @Input('opcionesCalendario')
+    set opcionesCalendario(value: any) {
+        this._opcionesCalendario = value;
+    }
+    get opcionesCalendario() {
+        return this._opcionesCalendario;
+    }
+
+    private _mostrarNoDisponibles;
+    @Input('mostrarNoDisponibles')
+    set mostrarNoDisponibles(value: any) {
+        this._mostrarNoDisponibles = value;
+    }
+    get mostrarNoDisponibles() {
+        return this._mostrarNoDisponibles;
+    }
 
     // Propiedades
-    @Output('agendaChanged') agendaChanged = new EventEmitter();
     @Input('fecha') fecha: Date;
+    @Input() _solicitudPrestacion: any;
     @Input('agenda')
     set agenda(value: IAgenda) {
         this._agenda = value;
@@ -44,6 +64,7 @@ export class CalendarioComponent {
     get estado(): String {
         return this._estado;
     }
+    @Output('agendaChanged') agendaChanged = new EventEmitter();
 
     /** Devuelve la primera agenda que encuentra de un día determinado */
     private agendaPorFecha(fecha: moment.Moment): IAgenda {
@@ -55,40 +76,57 @@ export class CalendarioComponent {
 
     /** Devuelve las agendas correspondientes a un día determinado */
     public agendasPorFecha(fecha: moment.Moment): IAgenda[] {
-        let ags = this.agendas.filter(
-            function (value) {
-                return (moment(fecha).isSame(value.horaInicio, 'day'));
-            }
-        );
+        let ags = this.agendas.filter((value) => {
+            return (moment(fecha).isSame(value.horaInicio, 'day'));
+        });
         return ags;
     }
+
     /** Regenera el calendario */
     private actualizar() {
+
         if (this.fecha && this.agendas) {
+
             let inicio = moment(this.fecha).startOf('month').startOf('week');
+            let cantidadSemanas = Math.ceil(moment(this.fecha).endOf('month').endOf('week').diff(moment(this.fecha).startOf('month').startOf('week'), 'weeks', true));
             this.diaSeleccionado = null;
             this.calendario = [];
-            for (let r = 1; r <= 5; r++) {
+
+            for (let r = 1; r <= cantidadSemanas; r++) {
                 let week = [];
                 this.calendario.push(week);
+
                 for (let c = 1; c <= 7; c++) {
-                    let ags = this.agendasPorFecha(inicio);
+
+                    let agendasPorFecha = this.agendasPorFecha(inicio);
                     let ag = null;
-                    if (ags.length > 1) {
-                        ag = ags[0];
+
+                    if (agendasPorFecha.length > 1) {
+
+                        ag = agendasPorFecha[0];
+
                         if (this.agenda) {
                             // Si hay una agenda seleccionada
-                            let i = ags.indexOf(this.agenda);
+                            let i = agendasPorFecha.indexOf(this.agenda);
                             if (i >= 0) {
-                                ag = ags[i];
+                                ag = agendasPorFecha[i];
                             }
                         }
                     }
-                    if (ags.length === 1) {
+
+                    if (agendasPorFecha.length === 1) {
                         ag = this.agendaPorFecha(inicio);
                     }
-                    let dia = new CalendarioDia(inicio.toDate(), ag);
-                    dia.cantidadAgendas = ags.length;
+
+                    let dia = new CalendarioDia(inicio.toDate(), ag, this._solicitudPrestacion);
+                    if (dia.estado === 'vacio' && this._solicitudPrestacion) {
+                        dia.cantidadAgendas = 0;
+                        dia.estado = 'vacio';
+                        dia.agenda = null;
+                    } else {
+                        dia.cantidadAgendas = agendasPorFecha.length;
+                    }
+                    dia.weekend = inicio.isoWeekday() >= 6;
                     week.push(dia);
 
                     // ¿Hay una agenda seleccionada?
