@@ -1,8 +1,8 @@
-export interface IBarrio{
-    id: String,
-    nombre: String,
+export interface IBarrio {
+    id: String;
+    nombre: String;
     localidad: {
             _id: String,
             nombre: String
-        }
+        };
 }
