@@ -1,8 +1,8 @@
 export interface ILocalidad {
-    id: String,
-    nombre: String,
+    id: String;
+    nombre: String;
     provincia: {
         id: String,
         nombre: String
-    }
+    };
 }
