@@ -1,7 +1,7 @@
 export interface IMatricula {
-      numero: Number,
-      descripcion: String,
-      activo: Boolean,
-      fechaInicio:Date,
-      fechaVencimiento: Date
+      numero: Number;
+      descripcion: String;
+      activo: Boolean;
+      fechaInicio: Date;
+      fechaVencimiento: Date;
 }
