@@ -3,22 +3,22 @@
  import { Sexo, Genero, EstadoCivil, tipoComunicacion } from './../utils/enumerados';
 
  export interface IProfesional {
-    id: String,
-    documento: String,
-    activo: Boolean,
-    nombre: String,
-    apellido: String,
+    id: String;
+    documento: String;
+    activo: Boolean;
+    nombre: String;
+    apellido: String;
     contacto: [{
         tipo: tipoComunicacion,
         valor: String,
         ranking: Number, // Specify preferred order of use (1 = highest) // Podemos usar el rank para guardar un historico de puntos de contacto (le restamos valor si no es actual???)
         ultimaActualizacion: Date,
         activo: Boolean
-    }],
-    sexo: Sexo,
-    genero: Genero, // identidad autopercibida
-    fechaNacimiento: Date, // Fecha Nacimiento
-    fechaFallecimiento: Date,
+    }];
+    sexo: Sexo;
+    genero: Genero; // identidad autopercibida
+    fechaNacimiento: Date; // Fecha Nacimiento
+    fechaFallecimiento: Date;
     direccion: [{
         valor: String,
         codigoPostal: String,
@@ -30,14 +30,14 @@
         },
         ultimaActualizacion: Date,
         activo: Boolean
-    }],
-    estadoCivil: EstadoCivil,
-    foto: String,
-    rol: String, //Ejemplo Jefe de Terapia intensiva
-    especialidad: [{ //El listado de sus especialidades
+    }];
+    estadoCivil: EstadoCivil;
+    foto: String;
+    rol: String; // Ejemplo Jefe de Terapia intensiva
+    especialidad: [{ // El listado de sus especialidades
         id: string,
         nombre: String
-    }],
+    }];
     matriculas: [{
         numero: Number,
         descripcion: String,
@@ -45,5 +45,5 @@
         fechaVencimiento: Date,
         activo: Boolean
 
-    }]
+    }];
  }
