@@ -1,8 +1,8 @@
-export interface IProvincia{
-    id: String,
-    nombre: String,
+export interface IProvincia {
+    id: String;
+    nombre: String;
     pais: {
             id: String,
             nombre: String
-        }
+        };
 }
