@@ -1,18 +1,18 @@
-export interface IUbicacion{
-    barrio : {
+export interface IUbicacion {
+    barrio: {
         id: String,
         nombre: String
-    }
+    };
     localidad: {
         id: String,
         nombre: String
-    },
+    };
     provincia: {
         id: String,
         nombre: String
-    },
+    };
     pais: {
         id: String,
         nombre: String
-    }
+    };
 }
