@@ -12,9 +12,21 @@ export interface IAgenda {
         nombre: String,
         apellido: String
     }];
-    espacioFisico: {
+    organizacion: {
         id: String,
         nombre: String
+    };
+    espacioFisico: {
+        id: String,
+        nombre: String,
+        servicio: {
+            id: String,
+            nombre: String
+        };
+        sector: {
+            id: String,
+            nombre: String
+        };
     };
     horaInicio: Date;
     horaFin: Date;
@@ -22,10 +34,17 @@ export interface IAgenda {
     bloques: IBloque[];
     estado: String;
     prePausada: String;
-    turnosDisponibles: number;
+    turnosDisponibles: number; // Virtual
+    turnosRestantesDelDia: number; // Virtual
+    turnosRestantesProgramados: number; // Virtual
+    turnosRestantesGestion: number; // Virtual
+    turnosRestantesProfesional: number; // Virtual
     estadosAgendas: String[];
-    organizacion: {
-        id: String,
-        nombre: String
-    }
+    nota: String;
+    nominalizada: Boolean;
+    avisos: [{
+        profenionalId: String,
+        estado: String,
+        fecha: Date
+    }];
 }
