@@ -2,18 +2,33 @@ import { IDireccion } from './../IDireccion';
 import { IContacto } from './../IContacto';
 export interface IEspacioFisico {
     id: string;
-    nombre: String;
-    descripcion: String;
-    edificio: {
-        id: String,
-        descripcion: String,
-        contacto: IContacto,
-        direccion: IDireccion
-    };
-    detalle: String;
-    activo: Boolean;
+    nombre: string;
+    descripcion: string;
+    detalle: string;
     organizacion: {
-        id: String,
-        nombre: String
+        id: string,
+        nombre: string
     };
+    edificio: {
+        id: string,
+        descripcion: string,
+        contacto: IContacto,
+        direccion: IDireccion,
+    };
+    sector: {
+        id: string,
+        nombre: string
+    };
+    servicio: {
+        id: string,
+        nombre: string
+    };
+    equipamiento: [{
+        conceptId: string;
+        term: string;
+        fsn: string;
+        semanticTag: string;
+    }];
+    activo: Boolean;
+    estado: string;
 }
