@@ -1,28 +1,7 @@
-
-// import { RupComponent } from './components/rup/rup.component';
-// import { PacienteService } from './services/paciente.service';
+// // ... MPI
 // import { HeaderPacienteComponent } from './components/paciente/headerPaciente.component';
-// // Angular
-// import { LOCALE_ID } from '@angular/core';
-// import { BrowserModule } from '@angular/platform-browser';
-// import { HttpModule } from '@angular/http';
-// import { FormsModule } from '@angular/forms';
-// import { NgModule } from '@angular/core';
+// // ... RUP
 
-// // Plex
-// import { PlexModule } from '@andes/plex/src/lib/module';
-// import { Plex } from '@andes/plex';
-
-// // Server
-// import { Server } from '@andes/shared';
-
-// // Routing
-// import { routing, appRoutingProviders } from './app.routing';
-
-// // RUP
-// import { EdadGestacionalComponent } from './components/rup/atomos/perinatales-nacimiento/edadGestacional.component';
-// import { ScoreApgarComponent } from './components/rup/atomos/perinatales-nacimiento/scoreApgar.component';
-// import { EdadGestacionalFetalComponent } from './components/rup/moleculas/edad-gestacional-fetal/edadGestacionalFetal.component';
 // import { PrestacionValidacionComponent } from './components/rup/ejecucion/prestacionValidacion.component';
 // import { PesoComponent } from './components/rup/atomos/peso.component';
 // import { SaturacionOxigenoComponent } from './components/rup/atomos/saturacionOxigeno.component';
@@ -32,119 +11,257 @@
 // import { FrecuenciaRespiratoriaComponent } from './components/rup/atomos/frecuenciaRespiratoria.component';
 // import { TensionSistolicaComponent } from './components/rup/atomos/tensionSistolica.component';
 // import { TensionDiastolicaComponent } from './components/rup/atomos/tensionDiastolica.component';
-// import { SignosVitalesComponent } from './components/rup/moleculas/signos-vitales/signosVitales.component';
-// import { TensionArterialComponent } from './components/rup/moleculas/tension-arterial/tensionArterial.component';
+// import { SignosVitalesComponent } from './components/rup/moleculas/signosVitales.component';
+// import { TensionArterialComponent } from './components/rup/moleculas/tensionArterial.component';
 // import { PrestacionEjecucionComponent } from './components/rup/ejecucion/prestacionEjecucion.component';
-// import { PuntoInicioComponent } from './components/rup/ejecucion/puntoInicio.component';
 // import { ResumenComponent } from './components/rup/ejecucion/resumen.component';
-// import { EvolucionTodosProblemasComponent } from './components/rup/problemas-paciente/evolucionTodosProblemas.component';
-// import { EvolucionProblemaComponent } from './components/rup/problemas-paciente/evolucionProblema.component';
-// import { TransformarProblemaComponent } from './components/rup/problemas-paciente/transformarProblema.component';
-// import { PruebaOtoemisionesAcusticasComponent } from './components/rup/atomos/pruebaOtoemisionesAcusticas.component';
+// // import { NuevoProblemaComponent } from './components/rup/problemas-paciente/nuevoProblema.component';
+// // import { EvolucionTodosProblemasComponent } from './components/rup/problemas-paciente/evolucionTodosProblemas.component';
+// // import { EvolucionProblemaComponent } from './components/rup/problemas-paciente/evolucionProblema.component';
+// // import { EnmendarProblemaComponent } from './components/rup/problemas-paciente/enmendarProblema.component';
+// // import { TransformarProblemaComponent } from './components/rup/problemas-paciente/transformarProblema.component';
+// // import { verProblemaComponent } from './components/rup/problemas-paciente/verProblema.component';
+// import { ConsultaMedicinaGeneralComponent } from './components/rup/moleculas/consultaMedicinaGeneral.component';
+// import { ObservacionesComponent } from './components/rup/atomos/observaciones.component';
+// import { PuntoInicioComponent } from './components/rup/ejecucion/puntoInicio.component';
+// import { EdadGestacionalComponent } from './components/rup/atomos/perinatales-nacimiento/edadGestacional.component';
+// import { ScoreApgarComponent } from './components/rup/atomos/perinatales-nacimiento/scoreApgar.component';
+// import { NacimientoComponent } from './components/rup/moleculas/nacimiento/nacimiento.component';
+// // import { EdadGestacionalFetalComponent } from './components/rup/moleculas/edad-gestacional-fetal/edadGestacionalFetal.component';
 // import { PesquisaNeonatalComponent } from './components/rup/atomos/pesquisaNeonatal.component';
-// import { PercentiloCircunferenciaCefalicaNinoComponent } from './components/rup/atomos/percentiloCircunferenciaCefalicaNino.component';
+// import { PruebaOtoemisionesAcusticasComponent } from './components/rup/atomos/pruebaOtoemisionesAcusticas.component';
 // import { PercentiloTallaComponent } from './components/rup/atomos/percentiloTalla.component';
+// import { PercentiloCircunferenciaCefalicaNinoComponent } from './components/rup/atomos/percentiloCircunferenciaCefalicaNino.component';
 // import { DesarrolloMotorComponent } from './components/rup/atomos/desarrolloMotor.component';
 // import { ActitudAnteLosCuidadosComponent } from './components/rup/atomos/actitudAnteLosCuidados.component';
 // import { ControlDeEsfinteresComponent } from './components/rup/atomos/controlDeEsfinteres.component';
 // import { DesarrolloIntelectualyJuegosComponent } from './components/rup/atomos/desarrolloIntelectualyJuegos.component';
 // import { EscalaDeDesarrolloComponent } from './components/rup/moleculas/escala-de-desarrollo/escalaDeDesarrollo.component';
-// import { FactoresDeRiesgoNinoSanoComponent } from './components/rup/atomos/factores-riesgo/factoresDeRiesgoNinoSano.component';
+// // import { FactoresDeRiesgoNinoSanoComponent } from './components/rup/atomos/factores-riesgo/factoresDeRiesgoNinoSano.component';
+// import { IndiceDeMasaCorporalComponent } from './components/rup/formulas/indiceDeMasaCorporal.component';
+// import { Formula } from './components/rup/core/formula.component';
+// import { RegistrosVisitasComponent } from './components/rup/moleculas/nino-sano/registro-visitas/registrosVisitas.component';
+// import { DatosPerinatalesComponent } from './components/rup/moleculas/nino-sano/datos-perinatales/datosPerinatales.component';
+// import { NinoSanoComponent } from './components/rup/moleculas/nino-sano/ninoSano.component';
+// import { Molecula } from './components/rup/core/molecula.component';
+// import { Atomo } from './components/rup/core/atomoComponent';
+// import { EcografiaComponent } from './components/rup/moleculas/ecografia.component';
+// import { InterConsultaComponent } from './components/rup/moleculas/interConsulta.component';
+// import { LaboratorioComponent } from './components/rup/moleculas/laboratorio.component';
+// import { RadiografiaComponent } from './components/rup/moleculas/radiografia.component';
+// import { TomografiaComponent } from './components/rup/moleculas/tomografia/tomografia.component';
+// import { MamografiaComponent } from './components/rup/moleculas/mamografia/mamografia.component';
+// import { EndoscopiaComponent } from './components/rup/moleculas/endoscopia/endoscopia.component';
+// import { TomaPapComponent } from './components/rup/moleculas/toma-pap/tomaPap.component';
+// import { TomaHpvComponent } from './components/rup/moleculas/toma-hpv/tomaHpv.component';
+// import { SangreOcultaMateriaFecalComponent } from './components/rup/moleculas/sangre-oculta-materia-fecal/sangreOcultaMateriaFecal.component';
+// import { BuscadorComponent } from './components/rup/ejecucion/buscador.component';
+// import { HudsBusquedaComponent } from './components/rup/ejecucion/hudsBusqueda.component';
+// import { SolicitudPrestacionDefaultComponent } from './components/rup/atomos/solicitudPrestacionDefault.component';
 
-// RUP
-import { ConsultaGeneralClinicaMedicaComponent } from './components/rup/moleculas/consulta-general-clinica-medica/consultaGeneralClinicaMedica.component';
-import { ObservacionesComponent } from './components/rup/atomos/observaciones.component';
-import { EdadGestacionalComponent } from './components/rup/atomos/perinatales-nacimiento/edadGestacional.component';
-import { ScoreApgarComponent } from './components/rup/atomos/perinatales-nacimiento/scoreApgar.component';
-import { EdadGestacionalFetalComponent } from './components/rup/moleculas/edad-gestacional-fetal/edadGestacionalFetal.component';
-import { PrestacionValidacionComponent } from './components/rup/ejecucion/prestacionValidacion.component';
-import { PesoComponent } from './components/rup/atomos/peso.component';
-import { SaturacionOxigenoComponent } from './components/rup/atomos/saturacionOxigeno.component';
-import { TallaComponent } from './components/rup/atomos/talla.component';
-import { TemperaturaComponent } from './components/rup/atomos/temperatura.component';
-import { FrecuenciaCardiacaComponent } from './components/rup/atomos/frecuenciaCardiaca.component';
-import { FrecuenciaRespiratoriaComponent } from './components/rup/atomos/frecuenciaRespiratoria.component';
-import { TensionSistolicaComponent } from './components/rup/atomos/tensionSistolica.component';
-import { TensionDiastolicaComponent } from './components/rup/atomos/tensionDiastolica.component';
-import { SignosVitalesComponent } from './components/rup/moleculas/signos-vitales/signosVitales.component';
-import { TensionArterialComponent } from './components/rup/moleculas/tension-arterial/tensionArterial.component';
-import { PrestacionEjecucionComponent } from './components/rup/ejecucion/prestacionEjecucion.component';
-import { PuntoInicioComponent } from './components/rup/ejecucion/puntoInicio.component';
-import { ResumenComponent } from './components/rup/ejecucion/resumen.component';
-import { EvolucionTodosProblemasComponent } from './components/rup/problemas-paciente/evolucionTodosProblemas.component';
-import { EvolucionProblemaComponent } from './components/rup/problemas-paciente/evolucionProblema.component';
-import { TransformarProblemaComponent } from './components/rup/problemas-paciente/transformarProblema.component';
-import { PruebaOtoemisionesAcusticasComponent } from './components/rup/atomos/pruebaOtoemisionesAcusticas.component';
-import { PesquisaNeonatalComponent } from './components/rup/atomos/pesquisaNeonatal.component';
-import { PercentiloCircunferenciaCefalicaNinoComponent } from './components/rup/atomos/percentiloCircunferenciaCefalicaNino.component';
-import { PercentiloTallaComponent } from './components/rup/atomos/percentiloTalla.component';
-import { DesarrolloMotorComponent } from './components/rup/atomos/desarrolloMotor.component';
-import { ActitudAnteLosCuidadosComponent } from './components/rup/atomos/actitudAnteLosCuidados.component';
-import { ControlDeEsfinteresComponent } from './components/rup/atomos/controlDeEsfinteres.component';
-import { DesarrolloIntelectualyJuegosComponent } from './components/rup/atomos/desarrolloIntelectualyJuegos.component';
-import { EscalaDeDesarrolloComponent } from './components/rup/moleculas/escala-de-desarrollo/escalaDeDesarrollo.component';
-import { FactoresDeRiesgoNinoSanoComponent } from './components/rup/atomos/factores-riesgo/factoresDeRiesgoNinoSano.component';
-import { IndiceDeMasaCorporalComponent } from './components/rup/moleculas/indice-de-masa-corporal/indiceDeMasaCorporal.component';
+// import { PrestacionCrearComponent } from './components/rup/ejecucion/prestacionCrear.component';
+// import { SnomedBuscarComponent } from './components/snomed/snomed-buscar.component';
+// import { EvolucionProblemaDefaultComponent } from './components/rup/atomos/evolucionProblemaDefault.component';
+// import { NuevaEvolucionProblemaComponent } from './components/rup/atomos/nuevaEvolucionProblema.component';
+// import { RefiereDefaultComponent } from './components/rup/atomos/refiereDefault.component';
 
-// ATOMO SOCIOECONOMICO
-import { ViviendaSituacionSocioEconomicaComponent } from './components/rup/moleculas/vivienda-situacion-socioeconomica-familiar/viviendaSituacionSocioEconomica.component';
-import { ViviendaResiduosComponent } from './components/rup/atomos/socio-economicos/viviendaResiduos.component';
-import { ViviendaPisoComponent } from './components/rup/atomos/socio-economicos/viviendaPiso.component';
-import { ViviendaNivelInstruccionComponent } from './components/rup/atomos/socio-economicos/viviendaNivelInstruccion.component';
-import { ViviendaFamiliaComponent } from './components/rup/atomos/socio-economicos/viviendaFamilia.component';
-import { ViviendaContaminantesComponent } from './components/rup/atomos/socio-economicos/viviendaContaminantes.component';
-import { ViviendaCombustionComponent } from './components/rup/atomos/socio-economicos/viviendaCombustion.component';
-import { ViviendaSostenEconomicoComponent } from './components/rup/atomos/socio-economicos/viviendaSostenEconomico.component';
-import { ViviendaAsistenciaEconomicaComponent } from './components/rup/atomos/socio-economicos/viviendaAsistenciaEconomica.component';
-import { ViviendaCondicionesAlojamientoComponent } from './components/rup/moleculas/vivienda-condiciones-y-alojamiento/viviendaCondicionesAlojamiento.component';
-// PERINATALES EMBARAZO
-import { PerinatalesEmbarazoNormalComponent } from './components/rup/atomos/perinatales-embarazo/perinatalesEmbarazoNormal.component';
-import { PerinatalesEmbarazoAnormalComponent } from './components/rup/atomos/perinatales-embarazo/perinatalesEmbarazoAnormal.component';
-import { PerinatalesNumeroGestaComponent } from './components/rup/atomos/perinatales-embarazo/perinatalesNumeroGesta.component';
-import { PerinatalesGestacionMultipleComponent } from './components/rup/atomos/perinatales-embarazo/perinatalesGestacionMultiple.component';
-// Perinatales Parto
-import { PartoViaVaginalComponent } from './components/rup/atomos/perinatales-parto/partoViaVaginal.component';
-import { PartoViaVaginalForcepsComponent } from './components/rup/atomos/perinatales-parto/partoViaVaginalForceps.component';
-import { PartoVaginalAsistidoExtractorVacioComponent } from './components/rup/atomos/perinatales-parto/partoVaginalAsistidoExtractorVacio.component';
-import { PartoCesareaComponent } from './components/rup/atomos/perinatales-parto/partoCesarea.component';
-import { HallazgoRelacionadoPartoComponent } from './components/rup/atomos/perinatales-parto/hallazgoRelacionadoParto.component';
-// NIÑO SANO
-import { OdontologiaComponent } from './components/rup/atomos/odontologia.component';
-import { EstadoNutricionalComponent } from './components/rup/atomos/estadoNutricional.component';
-import { HeaderPacienteComponent } from "./components/paciente/headerPaciente.component";
+// // ATOMO SOCIOECONOMICO
+// import { ViviendaSituacionSocioEconomicaComponent } from './components/rup/moleculas/vivienda-situacion-socioeconomica-familiar/viviendaSituacionSocioEconomica.component';
+// import { ViviendaResiduosComponent } from './components/rup/atomos/socio-economicos/viviendaResiduos.component';
+// import { ViviendaPisoComponent } from './components/rup/atomos/socio-economicos/viviendaPiso.component';
+// import { ViviendaNivelInstruccionComponent } from './components/rup/atomos/socio-economicos/viviendaNivelInstruccion.component';
+// import { ViviendaFamiliaComponent } from './components/rup/atomos/socio-economicos/viviendaFamilia.component';
+// import { ViviendaContaminantesComponent } from './components/rup/atomos/socio-economicos/viviendaContaminantes.component';
+// import { ViviendaCombustionComponent } from './components/rup/atomos/socio-economicos/viviendaCombustion.component';
+// import { ViviendaSostenEconomicoComponent } from './components/rup/atomos/socio-economicos/viviendaSostenEconomico.component';
+// import { ViviendaAsistenciaEconomicaComponent } from './components/rup/atomos/socio-economicos/viviendaAsistenciaEconomica.component';
+// import { ViviendaCondicionesAlojamientoComponent } from './components/rup/moleculas/vivienda-condiciones-y-alojamiento/viviendaCondicionesAlojamiento.component';
+// // PERINATALES EMBARAZO
+// import { PerinatalesEmbarazoNormalComponent } from './components/rup/atomos/perinatales-embarazo/perinatalesEmbarazoNormal.component';
+// import { PerinatalesEmbarazoAnormalComponent } from './components/rup/atomos/perinatales-embarazo/perinatalesEmbarazoAnormal.component';
+// import { PerinatalesNumeroGestaComponent } from './components/rup/atomos/perinatales-embarazo/perinatalesNumeroGesta.component';
+// import { PerinatalesGestacionMultipleComponent } from './components/rup/atomos/perinatales-embarazo/perinatalesGestacionMultiple.component';
+// // PERINATALES PARTO
+// import { PartoViaVaginalComponent } from './components/rup/atomos/perinatales-parto/partoViaVaginal.component';
+// import { PartoViaVaginalForcepsComponent } from './components/rup/atomos/perinatales-parto/partoViaVaginalForceps.component';
+// import { PartoVaginalAsistidoExtractorVacioComponent } from './components/rup/atomos/perinatales-parto/partoVaginalAsistidoExtractorVacio.component';
+// import { PartoCesareaComponent } from './components/rup/atomos/perinatales-parto/partoCesarea.component';
+// import { HallazgoRelacionadoPartoComponent } from './components/rup/atomos/perinatales-parto/hallazgoRelacionadoParto.component';
 
-// // Servicios RUP //
-// import { TipoPrestacionService } from './services/tipoPrestacion.service';
-// import { TipoProblemaService } from './services/rup/tipoProblema.service';
-// import { ProblemaPacienteService } from './services/rup/problemaPaciente.service';
-// import { PrestacionPacienteService } from './services/rup/prestacionPaciente.service';
+// // PRESCRIPCION MEDICAMENTOS
+// import { PrescripcionMedicamentoComponent } from './components/rup/moleculas/prescripcionMedicamento.component';
+// import { IndicacionProcedimientoComponent } from './components/rup/atomos/indicacionProcedimiento.component';
+// import { BuscadorMedicamentoSnomedComponent } from './components/rup/atomos/buscadorMedicamentoSnomed.component';
+// import { DosisComponent } from './components/rup/atomos/dosis.component';
+// import { RecetaComponent } from './components/rup/atomos/receta.component';
 
-// // Fin Servicios RUP //
+// // Autocitado
+// import { AutocitadoComponent } from './components/rup/atomos/autocitado.component';
 
 
+// // NIÑO SANO
+// // import { OdontologiaComponent } from './components/rup/atomos/odontologia.component';
+// import { EstadoNutricionalComponent } from './components/rup/atomos/estadoNutricional.component';
+
+// export const RUP_ELEMENTS: any[] = [
+//     { component: ActitudAnteLosCuidadosComponent, key: 'ActitudAnteLosCuidadosComponent' },
+//     { component: AutocitadoComponent, key: 'AutocitadoComponent' },
+//     { component: Atomo, key: 'Atomo' },
+//     { component: BuscadorComponent, key: 'BuscadorComponent' },
+//     { component: ConsultaMedicinaGeneralComponent, key: 'ConsultaMedicinaGeneralComponent' },
+//     { component: ControlDeEsfinteresComponent, key: 'ControlDeEsfinteresComponent' },
+//     { component: DatosPerinatalesComponent, key: 'DatosPerinatalesComponent' },
+//     { component: DesarrolloIntelectualyJuegosComponent, key: 'DesarrolloIntelectualyJuegosComponent' },
+//     { component: DesarrolloMotorComponent, key: 'DesarrolloMotorComponent' },
+//     { component: EcografiaComponent, key: 'EcografiaComponent' },
+//     { component: EdadGestacionalComponent, key: 'EdadGestacionalComponent' },
+//     // { component: EdadGestacionalFetalComponent, key: 'EdadGestacionalFetalComponent' },
+//     { component: EndoscopiaComponent, key: 'EndoscopiaComponent' },
+//     { component: EscalaDeDesarrolloComponent, key: 'EscalaDeDesarrolloComponent' },
+//     { component: EstadoNutricionalComponent, key: 'EstadoNutricionalComponent' },
+//     { component: EvolucionProblemaDefaultComponent, key: 'EvolucionProblemaDefaultComponent' },
+//     // { component: FactoresDeRiesgoNinoSanoComponent, key: 'FactoresDeRiesgoNinoSanoComponent' },
+//     { component: Formula, key: 'Formula' },
+//     { component: FrecuenciaCardiacaComponent, key: 'FrecuenciaCardiacaComponent' },
+//     { component: FrecuenciaRespiratoriaComponent, key: 'FrecuenciaRespiratoriaComponent' },
+//     { component: HallazgoRelacionadoPartoComponent, key: 'HallazgoRelacionadoPartoComponent' },
+//     { component: HeaderPacienteComponent, key: 'HeaderPacienteComponent' },
+//     { component: HudsBusquedaComponent, key: 'HudsBusquedaComponent' },
+//     { component: IndiceDeMasaCorporalComponent, key: 'IndiceDeMasaCorporalComponent' },
+//     { component: InterConsultaComponent, key: 'InterConsultaComponent' },
+//     { component: LaboratorioComponent, key: 'LaboratorioComponent' },
+//     { component: MamografiaComponent, key: 'MamografiaComponent' },
+//     { component: Molecula, key: 'Molecula' },
+//     { component: NacimientoComponent, key: 'NacimientoComponent' },
+//     { component: NinoSanoComponent, key: 'NinoSanoComponent' },
+//     { component: NuevaEvolucionProblemaComponent, key: 'NuevaEvolucionProblemaComponent' },
+//     { component: ObservacionesComponent, key: 'ObservacionesComponent' },
+//     // { component: OdontologiaComponent, key: 'OdontologiaComponent' },
+//     { component: PartoCesareaComponent, key: 'PartoCesareaComponent' },
+//     { component: PartoVaginalAsistidoExtractorVacioComponent, key: 'PartoVaginalAsistidoExtractorVacioComponent' },
+//     { component: PartoViaVaginalComponent, key: 'PartoViaVaginalComponent' },
+//     { component: PartoViaVaginalForcepsComponent, key: 'PartoViaVaginalForcepsComponent' },
+//     { component: PercentiloCircunferenciaCefalicaNinoComponent, key: 'PercentiloCircunferenciaCefalicaNinoComponent' },
+//     { component: PercentiloTallaComponent, key: 'PercentiloTallaComponent' },
+//     { component: PerinatalesEmbarazoAnormalComponent, key: 'PerinatalesEmbarazoAnormalComponent' },
+//     { component: PerinatalesEmbarazoNormalComponent, key: 'PerinatalesEmbarazoNormalComponent' },
+//     { component: PerinatalesGestacionMultipleComponent, key: 'PerinatalesGestacionMultipleComponent' },
+//     { component: PerinatalesNumeroGestaComponent, key: 'PerinatalesNumeroGestaComponent' },
+//     { component: PesoComponent, key: 'PesoComponent' },
+//     { component: PesquisaNeonatalComponent, key: 'PesquisaNeonatalComponent' },
+//     { component: PrestacionCrearComponent, key: 'PrestacionCrearComponent' },
+//     { component: PrestacionEjecucionComponent, key: 'PrestacionEjecucionComponent' },
+//     { component: PrestacionValidacionComponent, key: 'PrestacionValidacionComponent' },
+//     { component: PruebaOtoemisionesAcusticasComponent, key: 'PruebaOtoemisionesAcusticasComponent' },
+//     { component: PuntoInicioComponent, key: 'PuntoInicioComponent' },
+//     { component: RadiografiaComponent, key: 'RadiografiaComponent' },
+//     { component: RefiereDefaultComponent, key: 'RefiereDefaultComponent' },
+//     { component: RegistrosVisitasComponent, key: 'RegistrosVisitasComponent' },
+//     { component: ResumenComponent, key: 'ResumenComponent' },
+//     { component: SangreOcultaMateriaFecalComponent, key: 'SangreOcultaMateriaFecalComponent' },
+//     { component: SaturacionOxigenoComponent, key: 'SaturacionOxigenoComponent' },
+//     { component: ScoreApgarComponent, key: 'ScoreApgarComponent' },
+//     { component: SignosVitalesComponent, key: 'SignosVitalesComponent' },
+//     { component: SnomedBuscarComponent, key: 'SnomedBuscarComponent' },
+//     { component: SolicitudPrestacionDefaultComponent, key: 'SolicitudPrestacionDefaultComponent' },
+//     { component: TallaComponent, key: 'TallaComponent' },
+//     { component: TemperaturaComponent, key: 'TemperaturaComponent' },
+//     { component: TensionArterialComponent, key: 'TensionArterialComponent' },
+//     { component: TensionDiastolicaComponent, key: 'TensionDiastolicaComponent' },
+//     { component: TensionSistolicaComponent, key: 'TensionSistolicaComponent' },
+//     { component: TomaHpvComponent, key: 'TomaHpvComponent' },
+//     { component: TomaPapComponent, key: 'TomaPapComponent' },
+//     { component: TomografiaComponent, key: 'TomografiaComponent' },
+//     { component: TomografiaComponent, key: 'TomografiaComponent' },
+//     { component: ViviendaCombustionComponent, key: 'ViviendaCombustionComponent' },
+//     { component: ViviendaCondicionesAlojamientoComponent, key: 'ViviendaCondicionesAlojamientoComponent' },
+//     { component: ViviendaContaminantesComponent, key: 'ViviendaContaminantesComponent' },
+//     { component: ViviendaFamiliaComponent, key: 'ViviendaFamiliaComponent' },
+//     { component: ViviendaNivelInstruccionComponent, key: 'ViviendaNivelInstruccionComponent' },
+//     { component: ViviendaPisoComponent, key: 'ViviendaPisoComponent' },
+//     { component: ViviendaResiduosComponent, key: 'ViviendaResiduosComponent' },
+//     { component: ViviendaSituacionSocioEconomicaComponent, key: 'ViviendaSituacionSocioEconomicaComponent' },
+//     { component: ViviendaSostenEconomicoComponent, key: 'ViviendaSostenEconomicoComponent' },
+//     { component: PrescripcionMedicamentoComponent, key: 'PrescripcionMedicamentoComponent' },
+//     { component: IndicacionProcedimientoComponent, key: 'IndicacionProcedimientoComponent' },
+//     { component: BuscadorMedicamentoSnomedComponent, key: 'BuscadorMedicamentoSnomedComponent' },
+//     { component: DosisComponent, key: 'DosisComponent' },
+//     { component: RecetaComponent, key: 'RecetaComponent' }
+// ];
+
+// // let temp = RUP_ELEMENTS.map((i) => i.component);
 // export const RUP_COMPONENTS = [
-//     TensionArterialComponent,
-//     SignosVitalesComponent,
+//     AutocitadoComponent,
+//     BuscadorComponent,
+//     ConsultaMedicinaGeneralComponent,
+//     ControlDeEsfinteresComponent,
+//     DatosPerinatalesComponent,
+//     DesarrolloIntelectualyJuegosComponent,
+//     DesarrolloMotorComponent,
+//     EcografiaComponent,
+//     EndoscopiaComponent,
+//     EdadGestacionalComponent,
+//     ActitudAnteLosCuidadosComponent,
+//     Atomo,
+//     // EdadGestacionalFetalComponent,
+//     EscalaDeDesarrolloComponent,
+//     EstadoNutricionalComponent,
+//     EvolucionProblemaDefaultComponent,
+//     // FactoresDeRiesgoNinoSanoComponent,
+//     Formula,
 //     FrecuenciaCardiacaComponent,
 //     FrecuenciaRespiratoriaComponent,
+//     HallazgoRelacionadoPartoComponent,
+//     HeaderPacienteComponent,
+//     HudsBusquedaComponent,
+//     IndiceDeMasaCorporalComponent,
+//     InterConsultaComponent,
+//     LaboratorioComponent,
+//     MamografiaComponent,
+//     Molecula,
+//     NacimientoComponent,
+//     NinoSanoComponent,
+//     NuevaEvolucionProblemaComponent,
+//     ObservacionesComponent,
+//     // OdontologiaComponent,
+//     PartoCesareaComponent,
+//     PartoVaginalAsistidoExtractorVacioComponent,
+//     PartoViaVaginalComponent,
+//     PartoViaVaginalForcepsComponent,
+//     PercentiloCircunferenciaCefalicaNinoComponent,
+//     PercentiloTallaComponent,
+//     PerinatalesEmbarazoAnormalComponent,
+//     PerinatalesEmbarazoNormalComponent,
+//     PerinatalesGestacionMultipleComponent,
+//     PerinatalesNumeroGestaComponent,
 //     PesoComponent,
+//     PesquisaNeonatalComponent,
+//     PrestacionCrearComponent,
+//     PrestacionEjecucionComponent,
+//     PrestacionValidacionComponent,
+//     PruebaOtoemisionesAcusticasComponent,
+//     PuntoInicioComponent,
+//     RadiografiaComponent,
+//     RefiereDefaultComponent,
+//     RegistrosVisitasComponent,
+//     ResumenComponent,
+//     SangreOcultaMateriaFecalComponent,
 //     SaturacionOxigenoComponent,
+//     ScoreApgarComponent,
+//     SignosVitalesComponent,
+//     SnomedBuscarComponent,
+//     SolicitudPrestacionDefaultComponent,
 //     TallaComponent,
 //     TemperaturaComponent,
-//     TensionSistolicaComponent,
+//     TensionArterialComponent,
 //     TensionDiastolicaComponent,
-//     ObservacionesComponent,
-//     ConsultaGeneralClinicaMedicaComponent,
-//     EvolucionProblemaComponent,
-//     EvolucionTodosProblemasComponent,
-//     PuntoInicioComponent,
-//     PrestacionEjecucionComponent,
-//     ResumenComponent,
-//     HeaderPacienteComponent,
-//     PrestacionValidacionComponent,
-//     EdadGestacionalFetalComponent,
+//     TensionSistolicaComponent,
+//     TomaHpvComponent,
+//     TomaPapComponent,
+//     TomografiaComponent,
 //     ViviendaAsistenciaEconomicaComponent,
 //     ViviendaCombustionComponent,
+//     ViviendaCondicionesAlojamientoComponent,
 //     ViviendaContaminantesComponent,
 //     ViviendaFamiliaComponent,
 //     ViviendaNivelInstruccionComponent,
@@ -152,117 +269,9 @@ import { HeaderPacienteComponent } from "./components/paciente/headerPaciente.co
 //     ViviendaResiduosComponent,
 //     ViviendaSituacionSocioEconomicaComponent,
 //     ViviendaSostenEconomicoComponent,
-//     ViviendaCondicionesAlojamientoComponent,
-//     PerinatalesEmbarazoNormalComponent,
-//     PerinatalesEmbarazoAnormalComponent,
-//     PerinatalesNumeroGestaComponent,
-//     PerinatalesGestacionMultipleComponent,
-//     ScoreApgarComponent,
-//     EdadGestacionalComponent,
-//     PartoViaVaginalComponent,
-//     PartoVaginalAsistidoExtractorVacioComponent,
-//     PartoViaVaginalForcepsComponent,
-//     PartoCesareaComponent,
-//     HallazgoRelacionadoPartoComponent,
-//     TransformarProblemaComponent,
-//     PesquisaNeonatalComponent,
-//     PruebaOtoemisionesAcusticasComponent,
-//     PercentiloTallaComponent,
-//     PercentiloCircunferenciaCefalicaNinoComponent,
-//     DesarrolloMotorComponent,
-//     ActitudAnteLosCuidadosComponent,
-//     ControlDeEsfinteresComponent,
-//     DesarrolloIntelectualyJuegosComponent,
-//     EscalaDeDesarrolloComponent
-
-export const RUP_COMPONENTS = [
-    TensionArterialComponent,
-    SignosVitalesComponent,
-    FrecuenciaCardiacaComponent,
-    FrecuenciaRespiratoriaComponent,
-    PesoComponent,
-    SaturacionOxigenoComponent,
-    TallaComponent,
-    TemperaturaComponent,
-    TensionSistolicaComponent,
-    TensionDiastolicaComponent,
-    ObservacionesComponent,
-    ConsultaGeneralClinicaMedicaComponent,
-    EvolucionProblemaComponent,
-    EvolucionTodosProblemasComponent,
-    PuntoInicioComponent,
-    PrestacionEjecucionComponent,
-    ResumenComponent,
-    HeaderPacienteComponent,
-    PrestacionValidacionComponent,
-    EdadGestacionalFetalComponent,
-    ViviendaAsistenciaEconomicaComponent,
-    ViviendaCombustionComponent,
-    ViviendaContaminantesComponent,
-    ViviendaFamiliaComponent,
-    ViviendaNivelInstruccionComponent,
-    ViviendaPisoComponent,
-    ViviendaResiduosComponent,
-    ViviendaSituacionSocioEconomicaComponent,
-    ViviendaSostenEconomicoComponent,
-    ViviendaCondicionesAlojamientoComponent,
-    PerinatalesEmbarazoNormalComponent,
-    PerinatalesEmbarazoAnormalComponent,
-    PerinatalesNumeroGestaComponent,
-    PerinatalesGestacionMultipleComponent,
-    ScoreApgarComponent,
-    EdadGestacionalComponent,
-    PartoViaVaginalComponent,
-    PartoVaginalAsistidoExtractorVacioComponent,
-    PartoViaVaginalForcepsComponent,
-    PartoCesareaComponent,
-    HallazgoRelacionadoPartoComponent,
-    TransformarProblemaComponent,
-    PesquisaNeonatalComponent,
-    PruebaOtoemisionesAcusticasComponent,
-    PercentiloTallaComponent,
-    PercentiloCircunferenciaCefalicaNinoComponent,
-    DesarrolloMotorComponent,
-    ActitudAnteLosCuidadosComponent,
-    ControlDeEsfinteresComponent,
-    DesarrolloIntelectualyJuegosComponent,
-    EscalaDeDesarrolloComponent,
-    EstadoNutricionalComponent,
-    OdontologiaComponent,
-    IndiceDeMasaCorporalComponent,
-    FactoresDeRiesgoNinoSanoComponent
-    //     // NO VAPacienteComponent
-];
-
-// @NgModule({
-//     declarations: [
-//         RupComponent,
-//         ...RUP_COMPONENTS
-//     ],
-//     imports: [
-//         BrowserModule,
-//         FormsModule,
-//         HttpModule,
-//         PlexModule,
-//         routing
-//     ],
-//     exports: RUP_COMPONENTS,
-//     // [Andrrr] 2017-02-07: Requerido para poder crear componentes dinámicamente (Angular RC5-7)
-//     entryComponents: RUP_COMPONENTS,
-//     // bootstrap: [
-//     //     PuntoInicioComponent
-//     // ],
-//     providers: [
-//         { provide: LOCALE_ID, useValue: 'es-AR' },
-//         // PacienteService,
-//         // appRoutingProviders,
-//         // Plex,
-//         // Server,
-//         // PrestacionPacienteService,
-//         // ProblemaPacienteService,
-//         // TipoProblemaService,
-//         // TipoPrestacionService,
-
-//     ]
-// })
-// export class RupModule { debugger; }
+//     PrescripcionMedicamentoComponent,
+//     IndicacionProcedimientoComponent,
+//     BuscadorMedicamentoSnomedComponent,
+//     DosisComponent,
+//     RecetaComponent
+// ];
