@@ -1,3 +1,3 @@
-export function clonarObjeto(objeto: Object){
+export function clonarObjeto(objeto: Object) {
     return <any> JSON.parse(JSON.stringify(objeto));
 }
